Module google_content2::api

Structs

Account
Account data. After the creation of a new account it may take a few minutes before it is fully operational. The methods delete, insert, and update require the admin role.

AccountAddress
There is no detailed description.

AccountAdwordsLink
There is no detailed description.

AccountAuthinfoCall
Returns information about the authenticated user.

AccountBusinessInformation
There is no detailed description.

AccountClaimwebsiteCall
Claims the website of a Merchant Center sub-account.

AccountCustombatchCall
Retrieves, inserts, updates, and deletes multiple Merchant Center (sub-)accounts in a single request.

AccountCustomerService
There is no detailed description.

AccountDeleteCall
Deletes a Merchant Center sub-account.

AccountGetCall
Retrieves a Merchant Center account.

AccountGoogleMyBusinessLink
There is no detailed description.

AccountIdentifier
There is no detailed description.

AccountInsertCall
Creates a Merchant Center sub-account.

AccountLinkCall
Performs an action on a link between two Merchant Center accounts, namely accountId and linkedAccountId.

AccountListCall
Lists the sub-accounts in your Merchant Center account.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the ShoppingContent hub.

AccountStatus
The status of an account, i.e., information about its products, which is computed offline and not returned immediately at insertion time.

AccountStatusAccountLevelIssue
There is no detailed description.

AccountStatusDataQualityIssue
There is no detailed description.

AccountStatusExampleItem
There is no detailed description.

AccountStatusItemLevelIssue
There is no detailed description.

AccountStatusProducts
There is no detailed description.

AccountStatusStatistics
There is no detailed description.

AccountTax
The tax settings of a merchant account. All methods require the admin role.

AccountTaxTaxRule
Tax calculation rule to apply in a state or province (USA only).

AccountUpdateCall
Updates a Merchant Center account. Any fields that are not provided are deleted from the resource.

AccountUser
There is no detailed description.

AccountYouTubeChannelLink
There is no detailed description.

AccountsAuthInfoResponse
There is no detailed description.

AccountsClaimWebsiteResponse
There is no detailed description.

AccountsCustomBatchRequest
There is no detailed description.

AccountsCustomBatchRequestEntry
A batch entry encoding a single non-batch accounts request.

AccountsCustomBatchRequestEntryLinkRequest
There is no detailed description.

AccountsCustomBatchResponse
There is no detailed description.

AccountsCustomBatchResponseEntry
A batch entry encoding a single non-batch accounts response.

AccountsLinkRequest
There is no detailed description.

AccountsLinkResponse
There is no detailed description.

AccountsListResponse
There is no detailed description.

AccountstatuseCustombatchCall
Retrieves multiple Merchant Center account statuses in a single request.

AccountstatuseGetCall
Retrieves the status of a Merchant Center account. No itemLevelIssues are returned for multi-client accounts.

AccountstatuseListCall
Lists the statuses of the sub-accounts in your Merchant Center account.

AccountstatuseMethods
A builder providing access to all methods supported on accountstatuse resources. It is not used directly, but through the ShoppingContent hub.

AccountstatusesCustomBatchRequest
There is no detailed description.

AccountstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch accountstatuses request.

AccountstatusesCustomBatchResponse
There is no detailed description.

AccountstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch accountstatuses response.

AccountstatusesListResponse
There is no detailed description.

AccounttaxCustomBatchRequest
There is no detailed description.

AccounttaxCustomBatchRequestEntry
A batch entry encoding a single non-batch accounttax request.

AccounttaxCustomBatchResponse
There is no detailed description.

AccounttaxCustomBatchResponseEntry
A batch entry encoding a single non-batch accounttax response.

AccounttaxCustombatchCall
Retrieves and updates tax settings of multiple accounts in a single request.

AccounttaxGetCall
Retrieves the tax settings of the account.

AccounttaxListCall
Lists the tax settings of the sub-accounts in your Merchant Center account.

AccounttaxListResponse
There is no detailed description.

AccounttaxMethods
A builder providing access to all methods supported on accounttax resources. It is not used directly, but through the ShoppingContent hub.

AccounttaxUpdateCall
Updates the tax settings of the account. Any fields that are not provided are deleted from the resource.

Amount
There is no detailed description.

BusinessDayConfig
There is no detailed description.

CarrierRate
There is no detailed description.

CarriersCarrier
There is no detailed description.

CustomAttribute
There is no detailed description.

CustomGroup
There is no detailed description.

CustomerReturnReason
There is no detailed description.

CutoffTime
There is no detailed description.

Datafeed
Datafeed configuration data.

DatafeedCustombatchCall
Deletes, fetches, gets, inserts and updates multiple datafeeds in a single request.

DatafeedDeleteCall
Deletes a datafeed configuration from your Merchant Center account.

DatafeedFetchSchedule
The required fields vary based on the frequency of fetching. For a monthly fetch schedule, day_of_month and hour are required. For a weekly fetch schedule, weekday and hour are required. For a daily fetch schedule, only hour is required.

DatafeedFetchnowCall
Invokes a fetch for the datafeed in your Merchant Center account. If you need to call this method more than once per day, we recommend you use the Products service to update your product data.

DatafeedFormat
There is no detailed description.

DatafeedGetCall
Retrieves a datafeed configuration from your Merchant Center account.

DatafeedInsertCall
Registers a datafeed configuration with your Merchant Center account.

DatafeedListCall
Lists the configurations for datafeeds in your Merchant Center account.

DatafeedMethods
A builder providing access to all methods supported on datafeed resources. It is not used directly, but through the ShoppingContent hub.

DatafeedStatus
The status of a datafeed, i.e., the result of the last retrieval of the datafeed computed asynchronously when the feed processing is finished.

DatafeedStatusError
An error occurring in the feed, like "invalid price".

DatafeedStatusExample
An example occurrence for a particular error.

DatafeedTarget
There is no detailed description.

DatafeedUpdateCall
Updates a datafeed configuration of your Merchant Center account. Any fields that are not provided are deleted from the resource.

DatafeedsCustomBatchRequest
There is no detailed description.

DatafeedsCustomBatchRequestEntry
A batch entry encoding a single non-batch datafeeds request.

DatafeedsCustomBatchResponse
There is no detailed description.

DatafeedsCustomBatchResponseEntry
A batch entry encoding a single non-batch datafeeds response.

DatafeedsFetchNowResponse
There is no detailed description.

DatafeedsListResponse
There is no detailed description.

DatafeedstatuseCustombatchCall
Gets multiple Merchant Center datafeed statuses in a single request.

DatafeedstatuseGetCall
Retrieves the status of a datafeed from your Merchant Center account.

DatafeedstatuseListCall
Lists the statuses of the datafeeds in your Merchant Center account.

DatafeedstatuseMethods
A builder providing access to all methods supported on datafeedstatuse resources. It is not used directly, but through the ShoppingContent hub.

DatafeedstatusesCustomBatchRequest
There is no detailed description.

DatafeedstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch datafeedstatuses request.

DatafeedstatusesCustomBatchResponse
There is no detailed description.

DatafeedstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch datafeedstatuses response.

DatafeedstatusesListResponse
There is no detailed description.

DeliveryTime
There is no detailed description.

Error
An error returned by the API.

Errors
A list of errors returned by a failed batch entry.

GmbAccounts
There is no detailed description.

GmbAccountsGmbAccount
There is no detailed description.

Headers
A non-empty list of row or column headers for a table. Exactly one of prices, weights, numItems, postalCodeGroupNames, or location must be set.

HolidayCutoff
There is no detailed description.

HolidaysHoliday
There is no detailed description.

Installment
There is no detailed description.

Inventory
There is no detailed description.

InventoryCustomBatchRequest
There is no detailed description.

InventoryCustomBatchRequestEntry
A batch entry encoding a single non-batch inventory request.

InventoryCustomBatchResponse
There is no detailed description.

InventoryCustomBatchResponseEntry
A batch entry encoding a single non-batch inventory response.

InventoryCustombatchCall
Updates price and availability for multiple products or stores in a single request. This operation does not update the expiration date of the products.

InventoryMethods
A builder providing access to all methods supported on inventory resources. It is not used directly, but through the ShoppingContent hub.

InventoryPickup
There is no detailed description.

InventorySetCall
Updates price and availability of a product in your Merchant Center account.

InventorySetRequest
There is no detailed description.

InventorySetResponse
There is no detailed description.

InvoiceSummary
There is no detailed description.

InvoiceSummaryAdditionalChargeSummary
There is no detailed description.

LiaAboutPageSettings
There is no detailed description.

LiaCountrySettings
There is no detailed description.

LiaInventorySettings
There is no detailed description.

LiaOnDisplayToOrderSettings
There is no detailed description.

LiaPosDataProvider
There is no detailed description.

LiaSettings
Local Inventory ads (LIA) settings. All methods except listposdataproviders require the admin role.

LiasettingCustombatchCall
Retrieves and/or updates the LIA settings of multiple accounts in a single request.

LiasettingGetCall
Retrieves the LIA settings of the account.

LiasettingGetaccessiblegmbaccountCall
Retrieves the list of accessible Google My Business accounts.

LiasettingListCall
Lists the LIA settings of the sub-accounts in your Merchant Center account.

LiasettingListposdataproviderCall
Retrieves the list of POS data providers that have active settings for the all eiligible countries.

LiasettingMethods
A builder providing access to all methods supported on liasetting resources. It is not used directly, but through the ShoppingContent hub.

LiasettingRequestgmbaccesCall
Requests access to a specified Google My Business account.

LiasettingRequestinventoryverificationCall
Requests inventory validation for the specified country.

LiasettingSetinventoryverificationcontactCall
Sets the inventory verification contract for the specified country.

LiasettingSetposdataproviderCall
Sets the POS data provider for the specified country.

LiasettingUpdateCall
Updates the LIA settings of the account. Any fields that are not provided are deleted from the resource.

LiasettingsCustomBatchRequest
There is no detailed description.

LiasettingsCustomBatchRequestEntry
There is no detailed description.

LiasettingsCustomBatchResponse
There is no detailed description.

LiasettingsCustomBatchResponseEntry
There is no detailed description.

LiasettingsGetAccessibleGmbAccountsResponse
There is no detailed description.

LiasettingsListPosDataProvidersResponse
There is no detailed description.

LiasettingsListResponse
There is no detailed description.

LiasettingsRequestGmbAccessResponse
There is no detailed description.

LiasettingsRequestInventoryVerificationResponse
There is no detailed description.

LiasettingsSetInventoryVerificationContactResponse
There is no detailed description.

LiasettingsSetPosDataProviderResponse
There is no detailed description.

LocationIdSet
There is no detailed description.

LoyaltyPoints
There is no detailed description.

MerchantOrderReturn
Order return. Production access (all methods) requires the order manager role. Sandbox access does not.

MerchantOrderReturnItem
There is no detailed description.

MinimumOrderValueTable
There is no detailed description.

MinimumOrderValueTableStoreCodeSetWithMov
A list of store code sets sharing the same minimum order value. At least two sets are required and the last one must be empty, which signifies 'MOV for all other stores'. Each store code can only appear once across all the sets. All prices within a service must have the same currency.

Order
Order. Production access (all methods) requires the order manager role. Sandbox access does not.

OrderAcknowledgeCall
Marks an order as acknowledged.

OrderAddress
There is no detailed description.

OrderAdvancetestorderCall
Sandbox only. Moves a test order from state "inProgress" to state "pendingShipment".

OrderCancelCall
Cancels all line items in an order, making a full refund.

OrderCancellation
There is no detailed description.

OrderCancellineitemCall
Cancels a line item, making a full refund.

OrderCanceltestorderbycustomerCall
Sandbox only. Cancels a test order for customer-initiated cancellation.

OrderCreatetestorderCall
Sandbox only. Creates a test order.

OrderCreatetestreturnCall
Sandbox only. Creates a test return.

OrderCustombatchCall
Retrieves or modifies multiple orders in a single request.

OrderCustomer
There is no detailed description.

OrderCustomerMarketingRightsInfo
There is no detailed description.

OrderDeliveryDetails
There is no detailed description.

OrderGetCall
Retrieves an order from your Merchant Center account.

OrderGetbymerchantorderidCall
Retrieves an order using merchant order ID.

OrderGettestordertemplateCall
Sandbox only. Retrieves an order template that can be used to quickly create a new order in sandbox.

OrderInstorerefundlineitemCall
Deprecated. Notifies that item return and refund was handled directly by merchant outside of Google payments processing (e.g. cash refund done in store). Note: We recommend calling the returnrefundlineitem method to refund in-store returns. We will issue the refund directly to the customer. This helps to prevent possible differences arising between merchant and Google transaction records. We also recommend having the point of sale system communicate with Google to ensure that customers do not receive a double refund by first refunding via Google then via an in-store return.

OrderLegacyPromotion
There is no detailed description.

OrderLegacyPromotionBenefit
There is no detailed description.

OrderLineItem
There is no detailed description.

OrderLineItemProduct
There is no detailed description.

OrderLineItemProductFee
There is no detailed description.

OrderLineItemProductVariantAttribute
There is no detailed description.

OrderLineItemReturnInfo
There is no detailed description.

OrderLineItemShippingDetails
There is no detailed description.

OrderLineItemShippingDetailsMethod
There is no detailed description.

OrderListCall
Lists the orders in your Merchant Center account.

OrderMerchantProvidedAnnotation
There is no detailed description.

OrderMethods
A builder providing access to all methods supported on order resources. It is not used directly, but through the ShoppingContent hub.

OrderPaymentMethod
There is no detailed description.

OrderPickupDetails
There is no detailed description.

OrderPickupDetailsCollector
There is no detailed description.

OrderRefund
There is no detailed description.

OrderRefundCall
Deprecated, please use returnRefundLineItem instead.

OrderRejectreturnlineitemCall
Rejects return on an line item.

OrderReportDisbursement
Order disbursement. All methods require the payment analyst role.

OrderReportTransaction
There is no detailed description.

OrderReturn
There is no detailed description.

OrderReturnlineitemCall
Returns a line item.

OrderReturnrefundlineitemCall
Returns and refunds a line item. Note that this method can only be called on fully shipped orders. Please also note that the Orderreturns API is the preferred way to handle returns after you receive a return from a customer. You can use Orderreturns.list or Orderreturns.get to search for the return, and then use Orderreturns.processreturn to issue the refund. If the return cannot be found, then we recommend using this API to issue a refund.

OrderSetlineitemmetadataCall
Sets (or overrides if it already exists) merchant provided annotations in the form of key-value pairs. A common use case would be to supply us with additional structured information about a line item that cannot be provided via other methods. Submitted key-value pairs can be retrieved as part of the orders resource.

OrderShiplineitemCall
Marks line item(s) as shipped.

OrderShipment
There is no detailed description.

OrderShipmentLineItemShipment
There is no detailed description.

OrderShipmentScheduledDeliveryDetails
There is no detailed description.

OrderUpdatelineitemshippingdetailCall
Updates ship by and delivery by dates for a line item.

OrderUpdatemerchantorderidCall
Updates the merchant order ID for a given order.

OrderUpdateshipmentCall
Updates a shipment's status, carrier, and/or tracking ID.

OrderinvoiceCreatechargeinvoiceCall
Creates a charge invoice for a shipment group, and triggers a charge capture for orderinvoice enabled orders.

OrderinvoiceCreaterefundinvoiceCall
Creates a refund invoice for one or more shipment groups, and triggers a refund for orderinvoice enabled orders. This can only be used for line items that have previously been charged using createChargeInvoice. All amounts (except for the summary) are incremental with respect to the previous invoice.

OrderinvoiceMethods
A builder providing access to all methods supported on orderinvoice resources. It is not used directly, but through the ShoppingContent hub.

OrderinvoicesCreateChargeInvoiceRequest
There is no detailed description.

OrderinvoicesCreateChargeInvoiceResponse
There is no detailed description.

OrderinvoicesCreateRefundInvoiceRequest
There is no detailed description.

OrderinvoicesCreateRefundInvoiceResponse
There is no detailed description.

OrderinvoicesCustomBatchRequestEntryCreateRefundInvoiceRefundOption
There is no detailed description.

OrderinvoicesCustomBatchRequestEntryCreateRefundInvoiceReturnOption
There is no detailed description.

OrderreportListdisbursementCall
Retrieves a report for disbursements from your Merchant Center account.

OrderreportListtransactionCall
Retrieves a list of transactions for a disbursement from your Merchant Center account.

OrderreportMethods
A builder providing access to all methods supported on orderreport resources. It is not used directly, but through the ShoppingContent hub.

OrderreportsListDisbursementsResponse
There is no detailed description.

OrderreportsListTransactionsResponse
There is no detailed description.

OrderreturnGetCall
Retrieves an order return from your Merchant Center account.

OrderreturnListCall
Lists order returns in your Merchant Center account.

OrderreturnMethods
A builder providing access to all methods supported on orderreturn resources. It is not used directly, but through the ShoppingContent hub.

OrderreturnsListResponse
There is no detailed description.

OrdersAcknowledgeRequest
There is no detailed description.

OrdersAcknowledgeResponse
There is no detailed description.

OrdersAdvanceTestOrderResponse
There is no detailed description.

OrdersCancelLineItemRequest
There is no detailed description.

OrdersCancelLineItemResponse
There is no detailed description.

OrdersCancelRequest
There is no detailed description.

OrdersCancelResponse
There is no detailed description.

OrdersCancelTestOrderByCustomerRequest
There is no detailed description.

OrdersCancelTestOrderByCustomerResponse
There is no detailed description.

OrdersCreateTestOrderRequest
There is no detailed description.

OrdersCreateTestOrderResponse
There is no detailed description.

OrdersCreateTestReturnRequest
There is no detailed description.

OrdersCreateTestReturnResponse
There is no detailed description.

OrdersCustomBatchRequest
There is no detailed description.

OrdersCustomBatchRequestEntry
There is no detailed description.

OrdersCustomBatchRequestEntryCancel
There is no detailed description.

OrdersCustomBatchRequestEntryCancelLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryCreateTestReturnReturnItem
There is no detailed description.

OrdersCustomBatchRequestEntryInStoreRefundLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryRefund
There is no detailed description.

OrdersCustomBatchRequestEntryRejectReturnLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryReturnLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryReturnRefundLineItem
There is no detailed description.

OrdersCustomBatchRequestEntrySetLineItemMetadata
There is no detailed description.

OrdersCustomBatchRequestEntryShipLineItems
There is no detailed description.

OrdersCustomBatchRequestEntryShipLineItemsShipmentInfo
There is no detailed description.

OrdersCustomBatchRequestEntryUpdateLineItemShippingDetails
There is no detailed description.

OrdersCustomBatchRequestEntryUpdateShipment
There is no detailed description.

OrdersCustomBatchResponse
There is no detailed description.

OrdersCustomBatchResponseEntry
There is no detailed description.

OrdersGetByMerchantOrderIdResponse
There is no detailed description.

OrdersGetTestOrderTemplateResponse
There is no detailed description.

OrdersInStoreRefundLineItemRequest
There is no detailed description.

OrdersInStoreRefundLineItemResponse
There is no detailed description.

OrdersListResponse
There is no detailed description.

OrdersRefundRequest
There is no detailed description.

OrdersRefundResponse
There is no detailed description.

OrdersRejectReturnLineItemRequest
There is no detailed description.

OrdersRejectReturnLineItemResponse
There is no detailed description.

OrdersReturnLineItemRequest
There is no detailed description.

OrdersReturnLineItemResponse
There is no detailed description.

OrdersReturnRefundLineItemRequest
There is no detailed description.

OrdersReturnRefundLineItemResponse
There is no detailed description.

OrdersSetLineItemMetadataRequest
There is no detailed description.

OrdersSetLineItemMetadataResponse
There is no detailed description.

OrdersShipLineItemsRequest
There is no detailed description.

OrdersShipLineItemsResponse
There is no detailed description.

OrdersUpdateLineItemShippingDetailsRequest
There is no detailed description.

OrdersUpdateLineItemShippingDetailsResponse
There is no detailed description.

OrdersUpdateMerchantOrderIdRequest
There is no detailed description.

OrdersUpdateMerchantOrderIdResponse
There is no detailed description.

OrdersUpdateShipmentRequest
There is no detailed description.

OrdersUpdateShipmentResponse
There is no detailed description.

PickupCarrierService
There is no detailed description.

PickupServicesPickupService
There is no detailed description.

PoCustombatchCall
Batches multiple POS-related calls in a single request.

PoDeleteCall
Deletes a store for the given merchant.

PoGetCall
Retrieves information about the given store.

PoInsertCall
Creates a store for the given merchant.

PoInventoryCall
Submit inventory for the given merchant.

PoListCall
Lists the stores of the target merchant.

PoMethods
A builder providing access to all methods supported on po resources. It is not used directly, but through the ShoppingContent hub.

PoSaleCall
Submit a sale event for the given merchant.

PosCustomBatchRequest
There is no detailed description.

PosCustomBatchRequestEntry
There is no detailed description.

PosCustomBatchResponse
There is no detailed description.

PosCustomBatchResponseEntry
There is no detailed description.

PosDataProviders
There is no detailed description.

PosDataProvidersPosDataProvider
There is no detailed description.

PosInventory
The absolute quantity of an item available at the given store.

PosInventoryRequest
There is no detailed description.

PosInventoryResponse
There is no detailed description.

PosListResponse
There is no detailed description.

PosSale
The change of the available quantity of an item at the given store.

PosSaleRequest
There is no detailed description.

PosSaleResponse
There is no detailed description.

PosStore
Store resource.

PostalCodeGroup
There is no detailed description.

PostalCodeRange
There is no detailed description.

Price
There is no detailed description.

Product
Required product attributes are primarily defined by the products data specification. See the Products Data Specification Help Center article for information. Some attributes are country-specific, so make sure you select the appropriate country in the drop-down selector at the top of the page. Product data. After inserting, updating, or deleting a product, it may take several minutes before changes take effect.

ProductAmount
There is no detailed description.

ProductAspect
There is no detailed description.

ProductCustombatchCall
Retrieves, inserts, and deletes multiple products in a single request.

ProductDeleteCall
Deletes a product from your Merchant Center account.

ProductDestination
There is no detailed description.

ProductGetCall
Retrieves a product from your Merchant Center account.

ProductInsertCall
Uploads a product to your Merchant Center account. If an item with the same channel, contentLanguage, offerId, and targetCountry already exists, this method updates that entry.

ProductListCall
Lists the products in your Merchant Center account. The response might contain fewer items than specified by maxResults. Rely on nextPageToken to determine if there are more items to be requested.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the ShoppingContent hub.

ProductShipping
There is no detailed description.

ProductShippingDimension
There is no detailed description.

ProductShippingWeight
There is no detailed description.

ProductStatus
The status of a product, i.e., information about a product computed asynchronously.

ProductStatusDataQualityIssue
There is no detailed description.

ProductStatusDestinationStatus
There is no detailed description.

ProductStatusItemLevelIssue
There is no detailed description.

ProductTax
There is no detailed description.

ProductUnitPricingBaseMeasure
There is no detailed description.

ProductUnitPricingMeasure
There is no detailed description.

ProductsCustomBatchRequest
There is no detailed description.

ProductsCustomBatchRequestEntry
A batch entry encoding a single non-batch products request.

ProductsCustomBatchResponse
There is no detailed description.

ProductsCustomBatchResponseEntry
A batch entry encoding a single non-batch products response.

ProductsListResponse
There is no detailed description.

ProductstatuseCustombatchCall
Gets the statuses of multiple products in a single request.

ProductstatuseGetCall
Gets the status of a product from your Merchant Center account.

ProductstatuseListCall
Lists the statuses of the products in your Merchant Center account.

ProductstatuseMethods
A builder providing access to all methods supported on productstatuse resources. It is not used directly, but through the ShoppingContent hub.

ProductstatusesCustomBatchRequest
There is no detailed description.

ProductstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch productstatuses request.

ProductstatusesCustomBatchResponse
There is no detailed description.

ProductstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch productstatuses response.

ProductstatusesListResponse
There is no detailed description.

Promotion
There is no detailed description.

RateGroup
There is no detailed description.

RefundReason
There is no detailed description.

ReturnShipment
There is no detailed description.

Row
There is no detailed description.

Service
There is no detailed description.

ShipmentInvoice
There is no detailed description.

ShipmentInvoiceLineItemInvoice
There is no detailed description.

ShipmentTrackingInfo
There is no detailed description.

ShippingSettings
The merchant account's shipping settings. All methods except getsupportedcarriers and getsupportedholidays require the admin role.

ShippingsettingCustombatchCall
Retrieves and updates the shipping settings of multiple accounts in a single request.

ShippingsettingGetCall
Retrieves the shipping settings of the account.

ShippingsettingGetsupportedcarrierCall
Retrieves supported carriers and carrier services for an account.

ShippingsettingGetsupportedholidayCall
Retrieves supported holidays for an account.

ShippingsettingGetsupportedpickupserviceCall
Retrieves supported pickup services for an account.

ShippingsettingListCall
Lists the shipping settings of the sub-accounts in your Merchant Center account.

ShippingsettingMethods
A builder providing access to all methods supported on shippingsetting resources. It is not used directly, but through the ShoppingContent hub.

ShippingsettingUpdateCall
Updates the shipping settings of the account. Any fields that are not provided are deleted from the resource.

ShippingsettingsCustomBatchRequest
There is no detailed description.

ShippingsettingsCustomBatchRequestEntry
A batch entry encoding a single non-batch shippingsettings request.

ShippingsettingsCustomBatchResponse
There is no detailed description.

ShippingsettingsCustomBatchResponseEntry
A batch entry encoding a single non-batch shipping settings response.

ShippingsettingsGetSupportedCarriersResponse
There is no detailed description.

ShippingsettingsGetSupportedHolidaysResponse
There is no detailed description.

ShippingsettingsGetSupportedPickupServicesResponse
There is no detailed description.

ShippingsettingsListResponse
There is no detailed description.

ShoppingContent
Central instance to access all ShoppingContent related resource activities

Table
There is no detailed description.

TestOrder
There is no detailed description.

TestOrderCustomer
There is no detailed description.

TestOrderCustomerMarketingRightsInfo
There is no detailed description.

TestOrderLineItem
There is no detailed description.

TestOrderLineItemProduct
There is no detailed description.

TestOrderPaymentMethod
There is no detailed description.

TransitTable
There is no detailed description.

TransitTableTransitTimeRow
There is no detailed description.

TransitTableTransitTimeRowTransitTimeValue
There is no detailed description.

UnitInvoice
There is no detailed description.

UnitInvoiceAdditionalCharge
There is no detailed description.

UnitInvoiceTaxLine
There is no detailed description.

Value
The single value of a rate group or the value of a rate group table's cell. Exactly one of noShipping, flatRate, pricePercentage, carrierRateName, subtableName must be set.

Weight
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.