Struct google_content2::api::Product

pub struct Product {
    pub additional_image_links: Option<Vec<String>>,
    pub additional_product_types: Option<Vec<String>>,
    pub adult: Option<bool>,
    pub adwords_grouping: Option<String>,
    pub adwords_labels: Option<Vec<String>>,
    pub adwords_redirect: Option<String>,
    pub age_group: Option<String>,
    pub aspects: Option<Vec<ProductAspect>>,
    pub availability: Option<String>,
    pub availability_date: Option<String>,
    pub brand: Option<String>,
    pub canonical_link: Option<String>,
    pub channel: Option<String>,
    pub color: Option<String>,
    pub condition: Option<String>,
    pub content_language: Option<String>,
    pub cost_of_goods_sold: Option<Price>,
    pub custom_attributes: Option<Vec<CustomAttribute>>,
    pub custom_groups: Option<Vec<CustomGroup>>,
    pub custom_label0: Option<String>,
    pub custom_label1: Option<String>,
    pub custom_label2: Option<String>,
    pub custom_label3: Option<String>,
    pub custom_label4: Option<String>,
    pub description: Option<String>,
    pub destinations: Option<Vec<ProductDestination>>,
    pub display_ads_id: Option<String>,
    pub display_ads_link: Option<String>,
    pub display_ads_similar_ids: Option<Vec<String>>,
    pub display_ads_title: Option<String>,
    pub display_ads_value: Option<f64>,
    pub energy_efficiency_class: Option<String>,
    pub expiration_date: Option<String>,
    pub gender: Option<String>,
    pub google_product_category: Option<String>,
    pub gtin: Option<String>,
    pub id: Option<String>,
    pub identifier_exists: Option<bool>,
    pub image_link: Option<String>,
    pub installment: Option<Installment>,
    pub is_bundle: Option<bool>,
    pub item_group_id: Option<String>,
    pub kind: Option<String>,
    pub link: Option<String>,
    pub loyalty_points: Option<LoyaltyPoints>,
    pub material: Option<String>,
    pub max_energy_efficiency_class: Option<String>,
    pub max_handling_time: Option<String>,
    pub min_energy_efficiency_class: Option<String>,
    pub min_handling_time: Option<String>,
    pub mobile_link: Option<String>,
    pub mpn: Option<String>,
    pub multipack: Option<String>,
    pub offer_id: Option<String>,
    pub online_only: Option<bool>,
    pub pattern: Option<String>,
    pub price: Option<Price>,
    pub product_type: Option<String>,
    pub promotion_ids: Option<Vec<String>>,
    pub sale_price: Option<Price>,
    pub sale_price_effective_date: Option<String>,
    pub sell_on_google_quantity: Option<String>,
    pub shipping: Option<Vec<ProductShipping>>,
    pub shipping_height: Option<ProductShippingDimension>,
    pub shipping_label: Option<String>,
    pub shipping_length: Option<ProductShippingDimension>,
    pub shipping_weight: Option<ProductShippingWeight>,
    pub shipping_width: Option<ProductShippingDimension>,
    pub size_system: Option<String>,
    pub size_type: Option<String>,
    pub sizes: Option<Vec<String>>,
    pub source: Option<String>,
    pub target_country: Option<String>,
    pub taxes: Option<Vec<ProductTax>>,
    pub title: Option<String>,
    pub unit_pricing_base_measure: Option<ProductUnitPricingBaseMeasure>,
    pub unit_pricing_measure: Option<ProductUnitPricingMeasure>,
    pub validated_destinations: Option<Vec<String>>,
    pub warnings: Option<Vec<Error>>,
}

Required product attributes are primarily defined by the products data specification. See the Products Data Specification Help Center article for information. Some attributes are country-specific, so make sure you select the appropriate country in the drop-down selector at the top of the page. Product data. After inserting, updating, or deleting a product, it may take several minutes before changes take effect.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

custombatch products (none)

delete products (none)

get products (response)

insert products (request|response)

list products (none)

Fields
additional_image_links: Option<Vec<String>>

Additional URLs of images of the item.

additional_product_types: Option<Vec<String>>

Additional categories of the item (formatted as in products data specification).

adult: Option<bool>

Should be set to true if the item is targeted towards adults.

adwords_grouping: Option<String>

Used to group items in an arbitrary way. Only for CPA%, discouraged otherwise.

adwords_labels: Option<Vec<String>>

Similar to adwords_grouping, but only works on CPC.

adwords_redirect: Option<String>

Allows advertisers to override the item URL when the product is shown within the context of Product Ads.

age_group: Option<String>

Target age group of the item. Acceptable values are: - "adult" - "infant" - "kids" - "newborn" - "toddler" - "youngAdult"

aspects: Option<Vec<ProductAspect>>

Deprecated. Do not use.

availability: Option<String>

Availability status of the item. Acceptable values are: - "in stock" - "out of stock" - "preorder"

availability_date: Option<String>

The day a pre-ordered product becomes available for delivery, in ISO 8601 format.

brand: Option<String>

Brand of the item.

canonical_link: Option<String>

URL for the canonical version of your item's landing page.

channel: Option<String>

Required. The item's channel (online or local). Acceptable values are: - "local" - "online"

color: Option<String>

Color of the item.

condition: Option<String>

Condition or state of the item. Acceptable values are: - "new" - "refurbished" - "used"

content_language: Option<String>

Required. The two-letter ISO 639-1 language code for the item.

cost_of_goods_sold: Option<Price>

Cost of goods sold. Used for gross profit reporting.

custom_attributes: Option<Vec<CustomAttribute>>

A list of custom (merchant-provided) attributes. It can also be used for submitting any attribute of the feed specification in its generic form (e.g., { "name": "size type", "value": "regular" }). This is useful for submitting attributes not explicitly exposed by the API, such as additional attributes used for Buy on Google (formerly known as Shopping Actions).

custom_groups: Option<Vec<CustomGroup>>

A list of custom (merchant-provided) custom attribute groups.

custom_label0: Option<String>

Custom label 0 for custom grouping of items in a Shopping campaign.

custom_label1: Option<String>

Custom label 1 for custom grouping of items in a Shopping campaign.

custom_label2: Option<String>

Custom label 2 for custom grouping of items in a Shopping campaign.

custom_label3: Option<String>

Custom label 3 for custom grouping of items in a Shopping campaign.

custom_label4: Option<String>

Custom label 4 for custom grouping of items in a Shopping campaign.

description: Option<String>

Description of the item.

destinations: Option<Vec<ProductDestination>>

Specifies the intended destinations for the product.

display_ads_id: Option<String>

An identifier for an item for dynamic remarketing campaigns.

display_ads_link: Option<String>

URL directly to your item's landing page for dynamic remarketing campaigns.

display_ads_similar_ids: Option<Vec<String>>

Advertiser-specified recommendations.

display_ads_title: Option<String>

Title of an item for dynamic remarketing campaigns.

display_ads_value: Option<f64>

Offer margin for dynamic remarketing campaigns.

energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU. Acceptable values are: - "A" - "A+" - "A++" - "A+++" - "B" - "C" - "D" - "E" - "F" - "G"

expiration_date: Option<String>

Date on which the item should expire, as specified upon insertion, in ISO 8601 format. The actual expiration date in Google Shopping is exposed in productstatuses as googleExpirationDate and might be earlier if expirationDate is too far in the future.

gender: Option<String>

Target gender of the item. Acceptable values are: - "female" - "male" - "unisex"

google_product_category: Option<String>

Google's category of the item (see Google product taxonomy). When querying products, this field will contain the user provided value. There is currently no way to get back the auto assigned google product categories through the API.

gtin: Option<String>

Global Trade Item Number (GTIN) of the item.

id: Option<String>

The REST ID of the product. Content API methods that operate on products take this as their productId parameter. The REST ID for a product is of the form channel:contentLanguage: targetCountry: offerId.

identifier_exists: Option<bool>

False when the item does not have unique product identifiers appropriate to its category, such as GTIN, MPN, and brand. Required according to the Unique Product Identifier Rules for all target countries except for Canada.

image_link: Option<String>

URL of an image of the item.

installment: Option<Installment>

Number and amount of installments to pay for an item.

is_bundle: Option<bool>

Whether the item is a merchant-defined bundle. A bundle is a custom grouping of different products sold by a merchant for a single price.

item_group_id: Option<String>

Shared identifier for all variants of the same product.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "content#product"

link: Option<String>

URL directly linking to your item's page on your website.

loyalty_points: Option<LoyaltyPoints>

Loyalty points that users receive after purchasing the item. Japan only.

material: Option<String>

The material of which the item is made.

max_energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU. Acceptable values are: - "A" - "A+" - "A++" - "A+++" - "B" - "C" - "D" - "E" - "F" - "G"

max_handling_time: Option<String>

Maximal product handling time (in business days).

min_energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU. Acceptable values are: - "A" - "A+" - "A++" - "A+++" - "B" - "C" - "D" - "E" - "F" - "G"

min_handling_time: Option<String>

Minimal product handling time (in business days).

mobile_link: Option<String>

URL for the mobile-optimized version of your item's landing page.

mpn: Option<String>

Manufacturer Part Number (MPN) of the item.

multipack: Option<String>

The number of identical products in a merchant-defined multipack.

offer_id: Option<String>

Required. A unique identifier for the item. Leading and trailing whitespaces are stripped and multiple whitespaces are replaced by a single whitespace upon submission. Only valid unicode characters are accepted. See the products feed specification for details. Note: Content API methods that operate on products take the REST ID of the product, not this identifier.

online_only: Option<bool>

Deprecated.

pattern: Option<String>

The item's pattern (e.g. polka dots).

price: Option<Price>

Price of the item.

product_type: Option<String>

Your category of the item (formatted as in products data specification).

promotion_ids: Option<Vec<String>>

The unique ID of a promotion.

sale_price: Option<Price>

Advertised sale price of the item.

sale_price_effective_date: Option<String>

Date range during which the item is on sale (see products data specification ).

sell_on_google_quantity: Option<String>

The quantity of the product that is available for selling on Google. Supported only for online products.

shipping: Option<Vec<ProductShipping>>

Shipping rules.

shipping_height: Option<ProductShippingDimension>

Height of the item for shipping.

shipping_label: Option<String>

The shipping label of the product, used to group product in account-level shipping rules.

shipping_length: Option<ProductShippingDimension>

Length of the item for shipping.

shipping_weight: Option<ProductShippingWeight>

Weight of the item for shipping.

shipping_width: Option<ProductShippingDimension>

Width of the item for shipping.

size_system: Option<String>

System in which the size is specified. Recommended for apparel items. Acceptable values are: - "AU" - "BR" - "CN" - "DE" - "EU" - "FR" - "IT" - "JP" - "MEX" - "UK" - "US"

size_type: Option<String>

The cut of the item. Recommended for apparel items. Acceptable values are: - "big and tall" - "maternity" - "oversize" - "petite" - "plus" - "regular"

sizes: Option<Vec<String>>

Size of the item. Only one value is allowed. For variants with different sizes, insert a separate product for each size with the same itemGroupId value (see size definition).

source: Option<String>

The source of the offer, i.e., how the offer was created. Acceptable values are: - "api" - "crawl" - "feed"

target_country: Option<String>

Required. The CLDR territory code for the item.

taxes: Option<Vec<ProductTax>>

Tax information.

title: Option<String>

Title of the item.

unit_pricing_base_measure: Option<ProductUnitPricingBaseMeasure>

The preference of the denominator of the unit price.

unit_pricing_measure: Option<ProductUnitPricingMeasure>

The measure and dimension of an item.

validated_destinations: Option<Vec<String>>

Deprecated. The read-only list of intended destinations which passed validation.

warnings: Option<Vec<Error>>

Read-only warnings.