Struct google_content2::Account

pub struct Account {
    pub reviews_url: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub adult_content: Option<bool>,
    pub adwords_links: Option<Vec<AccountAdwordsLink>>,
    pub website_url: Option<String>,
    pub seller_id: Option<String>,
    pub id: Option<String>,
    pub users: Option<Vec<AccountUser>>,
}

Account data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

custombatch accounts (none)

patch accounts (request|response)

get accounts (response)

update accounts (request|response)

list accounts (none)

authinfo accounts (none)

delete accounts (none)

insert accounts (request|response)

Fields

reviews_url

URL for individual seller reviews, i.e., reviews for each child account.

kind

Identifies what kind of resource this is. Value: the fixed string "content#account".

name

Display name for the account.

adult_content

Indicates whether the merchant sells adult content.

adwords_links

List of linked AdWords accounts, active or pending approval. To create a new link request, add a new link with status active to the list. It will remain is state pending until approved or rejected in the AdWords interface. To delete an active link or to cancel a link request, remove it from the list.

website_url

The merchant's website.

seller_id

Client-specific, locally-unique, internal ID for the child account.

id

Merchant Center account ID.

users

Users with access to the account. Every account (except for subaccounts) must have at least one admin user.