Struct google_content2::AccountMethods

pub struct AccountMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the ShoppingContent hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_content2 as content2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2::ShoppingContent;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `authinfo(...)`, `custombatch(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

fn authinfo(&self) -> AccountAuthinfoCall<'a, C, A>

Create a builder to help you perform the following task:

Returns information about the authenticated user.

fn patch(&self, request: Account, merchant_id: &str, account_id: &str) -> AccountPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a Merchant Center account. This method supports patch semantics.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

accountId - The ID of the account.

fn get(&self, merchant_id: &str, account_id: &str) -> AccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a Merchant Center account.

Arguments

merchantId - The ID of the managing account.

accountId - The ID of the account.

fn update(&self, request: Account, merchant_id: &str, account_id: &str) -> AccountUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a Merchant Center account.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

accountId - The ID of the account.

fn list(&self, merchant_id: &str) -> AccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the sub-accounts in your Merchant Center account.

Arguments

merchantId - The ID of the managing account.

fn delete(&self, merchant_id: &str, account_id: &str) -> AccountDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a Merchant Center sub-account.

Arguments

merchantId - The ID of the managing account.

accountId - The ID of the account.

fn custombatch(&self, request: AccountsCustomBatchRequest) -> AccountCustombatchCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves, inserts, updates, and deletes multiple Merchant Center (sub-)accounts in a single request.

Arguments

request - No description provided.

fn insert(&self, request: Account, merchant_id: &str) -> AccountInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a Merchant Center sub-account.

Arguments

request - No description provided.

merchantId - The ID of the managing account.