Struct google_content2::AccountShippingShippingServiceCostRule

pub struct AccountShippingShippingServiceCostRule {
    pub calculation_method: Option<AccountShippingShippingServiceCalculationMethod>,
    pub condition: Option<AccountShippingCondition>,
    pub children: Option<Vec<AccountShippingShippingServiceCostRule>>,
}

Building block of the cost calculation decision tree. - The tree root should have no condition and no calculation method. - All the children must have a condition on the same dimension. The first child matching a condition is entered, therefore, price and weight conditions form contiguous intervals. - The last child of an element must have no condition and matches all elements not previously matched. - Children and calculation method are mutually exclusive, and exactly one of them must be defined; the root must only have children.

This type is not used in any activity, and only used as part of another schema.

Fields
calculation_method: Option<AccountShippingShippingServiceCalculationMethod>

Final calculation method to be used only in leaf nodes.

condition: Option<AccountShippingCondition>

Condition for this rule to be applicable. If no condition is specified, the rule acts as a catch-all.

children: Option<Vec<AccountShippingShippingServiceCostRule>>

Subsequent rules to be applied, only for inner nodes. The last child must not specify a condition and acts as a catch-all.