Struct google_content2::AccountsCustomBatchRequestEntry

pub struct AccountsCustomBatchRequestEntry {
    pub batch_id: Option<u32>,
    pub account: Option<Account>,
    pub account_id: Option<String>,
    pub merchant_id: Option<String>,
    pub method: Option<String>,
}

A batch entry encoding a single non-batch accounts request.

This type is not used in any activity, and only used as part of another schema.

Fields

batch_id

An entry ID, unique within the batch request.

account

The account to create or update. Only defined if the method is insert or update.

account_id

The ID of the account to get or delete. Only defined if the method is get or delete.

merchant_id

The ID of the managing account.

method

no description provided