Struct google_content2::AccountshippingMethods

pub struct AccountshippingMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on accountshipping resources. It is not used directly, but through the ShoppingContent hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_content2 as content2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2::ShoppingContent;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `custombatch(...)`, `get(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.accountshipping();

Methods

impl<'a, C, A> AccountshippingMethods<'a, C, A>

fn list(&self, merchant_id: &str) -> AccountshippingListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the shipping settings of the sub-accounts in your Merchant Center account.

Arguments

merchantId - The ID of the managing account.

fn patch(&self, request: AccountShipping, merchant_id: &str, account_id: &str) -> AccountshippingPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the shipping settings of the account. This method supports patch semantics.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

accountId - The ID of the account for which to get/update account shipping settings.

fn update(&self, request: AccountShipping, merchant_id: &str, account_id: &str) -> AccountshippingUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the shipping settings of the account.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

accountId - The ID of the account for which to get/update account shipping settings.

fn custombatch(&self, request: AccountshippingCustomBatchRequest) -> AccountshippingCustombatchCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves and updates the shipping settings of multiple accounts in a single request.

Arguments

request - No description provided.

fn get(&self, merchant_id: &str, account_id: &str) -> AccountshippingGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the shipping settings of the account.

Arguments

merchantId - The ID of the managing account.

accountId - The ID of the account for which to get/update account shipping settings.