Struct google_content2::DatafeedFetchSchedule

pub struct DatafeedFetchSchedule {
    pub username: Option<String>,
    pub password: Option<String>,
    pub weekday: Option<String>,
    pub hour: Option<u32>,
    pub time_zone: Option<String>,
    pub fetch_url: Option<String>,
    pub day_of_month: Option<u32>,
}

The required fields vary based on the frequency of fetching. For a monthly fetch schedule, day_of_month and hour are required. For a weekly fetch schedule, weekday and hour are required. For a daily fetch schedule, only hour is required.

This type is not used in any activity, and only used as part of another schema.

Fields

username

An optional user name for fetch_url.

password

An optional password for fetch_url.

weekday

The day of the week the feed file should be fetched.

hour

The hour of the day the feed file should be fetched (0-23).

time_zone

Time zone used for schedule. UTC by default. E.g., "America/Los_Angeles".

fetch_url

The URL where the feed file can be fetched. Google Merchant Center will support automatic scheduled uploads using the HTTP, HTTPS, FTP, or SFTP protocols, so the value will need to be a valid link using one of those four protocols.

day_of_month

The day of the month the feed file should be fetched (1-31).