Struct google_content2::DatafeedStatus

pub struct DatafeedStatus {
    pub kind: Option<String>,
    pub errors: Option<Vec<DatafeedStatusError>>,
    pub warnings: Option<Vec<DatafeedStatusError>>,
    pub items_total: Option<String>,
    pub processing_status: Option<String>,
    pub last_upload_date: Option<String>,
    pub items_valid: Option<String>,
    pub datafeed_id: Option<String>,
}

The status of a datafeed, i.e., the result of the last retrieval of the datafeed computed asynchronously when the feed processing is finished.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get datafeedstatuses (response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "content#datafeedStatus".

errors

The list of errors occurring in the feed.

warnings

The list of errors occurring in the feed.

items_total

The number of items in the feed that were processed.

processing_status

The processing status of the feed.

last_upload_date

The last date at which the feed was uploaded.

items_valid

The number of items in the feed that were valid.

datafeed_id

The ID of the feed for which the status is reported.