Struct google_content2::Inventory

pub struct Inventory {
    pub kind: Option<String>,
    pub sale_price: Option<Price>,
    pub sale_price_effective_date: Option<String>,
    pub price: Option<Price>,
    pub availability: Option<String>,
    pub quantity: Option<u32>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "content#inventory".

sale_price

The sale price of the product. Mandatory if sale_price_effective_date is defined.

sale_price_effective_date

A date range represented by a pair of ISO 8601 dates separated by a space, comma, or slash. Both dates might be specified as 'null' if undecided.

price

The price of the product.

availability

The availability of the product.

quantity

The quantity of the product. Must be equal to or greater than zero. Supported only for local products.