Struct google_content2::InventoryMethods

pub struct InventoryMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on inventory resources. It is not used directly, but through the ShoppingContent hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_content2 as content2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2::ShoppingContent;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `custombatch(...)` and `set(...)`
// to build up your call.
let rb = hub.inventory();

Methods

impl<'a, C, A> InventoryMethods<'a, C, A>

fn set(&self, request: InventorySetRequest, merchant_id: &str, store_code: &str, product_id: &str) -> InventorySetCall<'a, C, A>

Create a builder to help you perform the following task:

Updates price and availability of a product in your Merchant Center account. This operation does not update the expiration date of the product.

Arguments

request - No description provided.

merchantId - The ID of the managing account.

storeCode - The code of the store for which to update price and availability. Use online to update price and availability of an online product.

productId - The ID of the product for which to update price and availability.

fn custombatch(&self, request: InventoryCustomBatchRequest) -> InventoryCustombatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates price and availability for multiple products or stores in a single request. This operation does not update the expiration date of the products.

Arguments

request - No description provided.