Struct google_content2::Order

pub struct Order {
    pub shipments: Option<Vec<OrderShipment>>,
    pub customer: Option<OrderCustomer>,
    pub payment_status: Option<String>,
    pub net_amount: Option<Price>,
    pub id: Option<String>,
    pub promotions: Option<Vec<OrderPromotion>>,
    pub kind: Option<String>,
    pub line_items: Option<Vec<OrderLineItem>>,
    pub refunds: Option<Vec<OrderRefund>>,
    pub acknowledged: Option<bool>,
    pub shipping_cost: Option<Price>,
    pub merchant_id: Option<String>,
    pub merchant_order_id: Option<String>,
    pub shipping_option: Option<String>,
    pub status: Option<String>,
    pub delivery_details: Option<OrderDeliveryDetails>,
    pub placed_date: Option<String>,
    pub shipping_cost_tax: Option<Price>,
    pub payment_method: Option<OrderPaymentMethod>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get orders (response)

list orders (none)

updateshipment orders (none)

advancetestorder orders (none)

updatemerchantorderid orders (none)

returnlineitem orders (none)

gettestordertemplate orders (none)

createtestorder orders (none)

refund orders (none)

custombatch orders (none)

cancellineitem orders (none)

getbymerchantorderid orders (none)

acknowledge orders (none)

cancel orders (none)

shiplineitems orders (none)

Fields

shipments

Shipments of the order.

customer

The details of the customer who placed the order.

payment_status

The status of the payment.

net_amount

The net amount for the order. For example, if an order was originally for a grand total of $100 and a refund was issued for $20, the net amount will be $80.

id

The REST id of the order. Globally unique.

promotions

The details of the merchant provided promotions applied to the order. More details about the program are here.

kind

Identifies what kind of resource this is. Value: the fixed string "content#order".

line_items

Line items that are ordered.

refunds

Refunds for the order.

acknowledged

Whether the order was acknowledged.

shipping_cost

The total cost of shipping for all items.

merchant_id

no description provided

merchant_order_id

Merchant-provided id of the order.

shipping_option

The requested shipping option.

status

The status of the order.

delivery_details

The details for the delivery.

placed_date

The date when the order was placed, in ISO 8601 format.

shipping_cost_tax

The tax for the total shipping cost.

payment_method

The details of the payment method.