Struct google_content2::Product

pub struct Product {
    pub max_handling_time: Option<String>,
    pub color: Option<String>,
    pub adwords_labels: Option<Vec<String>>,
    pub item_group_id: Option<String>,
    pub expiration_date: Option<String>,
    pub availability: Option<String>,
    pub adwords_grouping: Option<String>,
    pub destinations: Option<Vec<ProductDestination>>,
    pub target_country: Option<String>,
    pub custom_label4: Option<String>,
    pub source: Option<String>,
    pub additional_product_types: Option<Vec<String>>,
    pub min_energy_efficiency_class: Option<String>,
    pub online_only: Option<bool>,
    pub mobile_link: Option<String>,
    pub promotion_ids: Option<Vec<String>>,
    pub aspects: Option<Vec<ProductAspect>>,
    pub shipping_height: Option<ProductShippingDimension>,
    pub content_language: Option<String>,
    pub mpn: Option<String>,
    pub sale_price_effective_date: Option<String>,
    pub brand: Option<String>,
    pub cost_of_goods_sold: Option<Price>,
    pub link: Option<String>,
    pub adwords_redirect: Option<String>,
    pub energy_efficiency_class: Option<String>,
    pub size_system: Option<String>,
    pub size_type: Option<String>,
    pub custom_label3: Option<String>,
    pub custom_label2: Option<String>,
    pub condition: Option<String>,
    pub custom_label0: Option<String>,
    pub kind: Option<String>,
    pub installment: Option<Installment>,
    pub sizes: Option<Vec<String>>,
    pub gender: Option<String>,
    pub taxes: Option<Vec<ProductTax>>,
    pub max_energy_efficiency_class: Option<String>,
    pub identifier_exists: Option<bool>,
    pub sale_price: Option<Price>,
    pub product_type: Option<String>,
    pub display_ads_similar_ids: Option<Vec<String>>,
    pub custom_label1: Option<String>,
    pub age_group: Option<String>,
    pub display_ads_title: Option<String>,
    pub additional_image_links: Option<Vec<String>>,
    pub offer_id: Option<String>,
    pub min_handling_time: Option<String>,
    pub gtin: Option<String>,
    pub shipping_label: Option<String>,
    pub google_product_category: Option<String>,
    pub multipack: Option<String>,
    pub display_ads_id: Option<String>,
    pub display_ads_value: Option<f64>,
    pub id: Option<String>,
    pub shipping_width: Option<ProductShippingDimension>,
    pub pattern: Option<String>,
    pub unit_pricing_measure: Option<ProductUnitPricingMeasure>,
    pub custom_attributes: Option<Vec<CustomAttribute>>,
    pub availability_date: Option<String>,
    pub is_bundle: Option<bool>,
    pub channel: Option<String>,
    pub material: Option<String>,
    pub description: Option<String>,
    pub warnings: Option<Vec<ErrorType>>,
    pub price: Option<Price>,
    pub shipping_length: Option<ProductShippingDimension>,
    pub sell_on_google_quantity: Option<String>,
    pub display_ads_link: Option<String>,
    pub validated_destinations: Option<Vec<String>>,
    pub custom_groups: Option<Vec<CustomGroup>>,
    pub loyalty_points: Option<LoyaltyPoints>,
    pub unit_pricing_base_measure: Option<ProductUnitPricingBaseMeasure>,
    pub title: Option<String>,
    pub shipping: Option<Vec<ProductShipping>>,
    pub shipping_weight: Option<ProductShippingWeight>,
    pub image_link: Option<String>,
    pub adult: Option<bool>,
}

Product data. After inserting, updating, or deleting a product, it may take several minutes before changes take effect.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete products (none)

custombatch products (none)

get products (response)

list products (none)

insert products (request|response)

Fields
max_handling_time: Option<String>

Maximal product handling time (in business days).

color: Option<String>

Color of the item.

adwords_labels: Option<Vec<String>>

Similar to adwords_grouping, but only works on CPC.

item_group_id: Option<String>

Shared identifier for all variants of the same product.

expiration_date: Option<String>

Date on which the item should expire, as specified upon insertion, in ISO 8601 format. The actual expiration date in Google Shopping is exposed in productstatuses as googleExpirationDate and might be earlier if expirationDate is too far in the future.

availability: Option<String>

Availability status of the item.

adwords_grouping: Option<String>

Used to group items in an arbitrary way. Only for CPA%, discouraged otherwise.

destinations: Option<Vec<ProductDestination>>

Specifies the intended destinations for the product.

target_country: Option<String>

The CLDR territory code for the item.

custom_label4: Option<String>

Custom label 4 for custom grouping of items in a Shopping campaign.

source: Option<String>

The source of the offer, i.e., how the offer was created.

additional_product_types: Option<Vec<String>>

Additional categories of the item (formatted as in products data specification).

min_energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU.

online_only: Option<bool>

Deprecated. Whether an item is available for purchase only online.

mobile_link: Option<String>

Link to a mobile-optimized version of the landing page.

promotion_ids: Option<Vec<String>>

The unique ID of a promotion.

aspects: Option<Vec<ProductAspect>>

Deprecated. Do not use.

shipping_height: Option<ProductShippingDimension>

Height of the item for shipping.

content_language: Option<String>

The two-letter ISO 639-1 language code for the item.

mpn: Option<String>

Manufacturer Part Number (MPN) of the item.

sale_price_effective_date: Option<String>

Date range during which the item is on sale (see products data specification).

brand: Option<String>

Brand of the item.

cost_of_goods_sold: Option<Price>

Cost of goods sold. Used for gross profit reporting.

link: Option<String>

URL directly linking to your item's page on your website.

adwords_redirect: Option<String>

Allows advertisers to override the item URL when the product is shown within the context of Product Ads.

energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU.

size_system: Option<String>

System in which the size is specified. Recommended for apparel items.

size_type: Option<String>

The cut of the item. Recommended for apparel items.

custom_label3: Option<String>

Custom label 3 for custom grouping of items in a Shopping campaign.

custom_label2: Option<String>

Custom label 2 for custom grouping of items in a Shopping campaign.

condition: Option<String>

Condition or state of the item.

custom_label0: Option<String>

Custom label 0 for custom grouping of items in a Shopping campaign.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "content#product".

installment: Option<Installment>

Number and amount of installments to pay for an item. Brazil only.

sizes: Option<Vec<String>>

Size of the item.

gender: Option<String>

Target gender of the item.

taxes: Option<Vec<ProductTax>>

Tax information.

max_energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU.

identifier_exists: Option<bool>

False when the item does not have unique product identifiers appropriate to its category, such as GTIN, MPN, and brand. Required according to the Unique Product Identifier Rules for all target countries except for Canada.

sale_price: Option<Price>

Advertised sale price of the item.

product_type: Option<String>

Your category of the item (formatted as in products data specification).

display_ads_similar_ids: Option<Vec<String>>

Advertiser-specified recommendations.

custom_label1: Option<String>

Custom label 1 for custom grouping of items in a Shopping campaign.

age_group: Option<String>

Target age group of the item.

display_ads_title: Option<String>

Title of an item for dynamic remarketing campaigns.

additional_image_links: Option<Vec<String>>

Additional URLs of images of the item.

offer_id: Option<String>

A unique identifier for the item. Leading and trailing whitespaces are stripped and multiple whitespaces are replaced by a single whitespace upon submission. Only valid unicode characters are accepted. See the products feed specification for details. Note: Content API methods that operate on products take the REST ID of the product, not this identifier.

min_handling_time: Option<String>

Minimal product handling time (in business days).

gtin: Option<String>

Global Trade Item Number (GTIN) of the item.

shipping_label: Option<String>

The shipping label of the product, used to group product in account-level shipping rules.

google_product_category: Option<String>

Google's category of the item (see Google product taxonomy).

multipack: Option<String>

The number of identical products in a merchant-defined multipack.

display_ads_id: Option<String>

An identifier for an item for dynamic remarketing campaigns.

display_ads_value: Option<f64>

Offer margin for dynamic remarketing campaigns.

id: Option<String>

The REST ID of the product. Content API methods that operate on products take this as their productId parameter. The REST ID for a product is of the form channel:contentLanguage:targetCountry:offerId.

shipping_width: Option<ProductShippingDimension>

Width of the item for shipping.

pattern: Option<String>

The item's pattern (e.g. polka dots).

unit_pricing_measure: Option<ProductUnitPricingMeasure>

The measure and dimension of an item.

custom_attributes: Option<Vec<CustomAttribute>>

A list of custom (merchant-provided) attributes. It can also be used for submitting any attribute of the feed specification in its generic form (e.g., { "name": "size type", "value": "regular" }). This is useful for submitting attributes not explicitly exposed by the API.

availability_date: Option<String>

The day a pre-ordered product becomes available for delivery, in ISO 8601 format.

is_bundle: Option<bool>

Whether the item is a merchant-defined bundle. A bundle is a custom grouping of different products sold by a merchant for a single price.

channel: Option<String>

The item's channel (online or local).

material: Option<String>

The material of which the item is made.

description: Option<String>

Description of the item.

warnings: Option<Vec<ErrorType>>

Read-only warnings.

price: Option<Price>

Price of the item.

shipping_length: Option<ProductShippingDimension>

Length of the item for shipping.

sell_on_google_quantity: Option<String>

The quantity of the product that is available for selling on Google. Supported only for online products.

display_ads_link: Option<String>

URL directly to your item's landing page for dynamic remarketing campaigns.

validated_destinations: Option<Vec<String>>

Deprecated. The read-only list of intended destinations which passed validation.

custom_groups: Option<Vec<CustomGroup>>

A list of custom (merchant-provided) custom attribute groups.

loyalty_points: Option<LoyaltyPoints>

Loyalty points that users receive after purchasing the item. Japan only.

unit_pricing_base_measure: Option<ProductUnitPricingBaseMeasure>

The preference of the denominator of the unit price.

title: Option<String>

Title of the item.

shipping: Option<Vec<ProductShipping>>

Shipping rules.

shipping_weight: Option<ProductShippingWeight>

Weight of the item for shipping.

image_link: Option<String>

URL of an image of the item.

adult: Option<bool>

Set to true if the item is targeted towards adults.