Struct google_content2::ProductShipping

pub struct ProductShipping {
    pub location_id: Option<String>,
    pub service: Option<String>,
    pub postal_code: Option<String>,
    pub country: Option<String>,
    pub price: Option<Price>,
    pub region: Option<String>,
    pub location_group_name: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
location_id: Option<String>

The numeric id of a location that the shipping rate applies to as defined in the AdWords API.

service: Option<String>

A free-form description of the service class or delivery speed.

postal_code: Option<String>

The postal code range that the shipping rate applies to, represented by a postal code, a postal code prefix followed by a * wildcard, a range between two postal codes or two postal code prefixes of equal length.

country: Option<String>

The CLDR territory code of the country to which an item will ship.

price: Option<Price>

Fixed shipping price, represented as a number.

region: Option<String>

The geographic region to which a shipping rate applies (e.g. zip code).

location_group_name: Option<String>

The location where the shipping is applicable, represented by a location group name.