Struct google_content2::ProductStatus

pub struct ProductStatus {
    pub google_expiration_date: Option<String>,
    pub destination_statuses: Option<Vec<ProductStatusDestinationStatus>>,
    pub title: Option<String>,
    pub kind: Option<String>,
    pub last_update_date: Option<String>,
    pub link: Option<String>,
    pub creation_date: Option<String>,
    pub data_quality_issues: Option<Vec<ProductStatusDataQualityIssue>>,
    pub product_id: Option<String>,
}

The status of a product, i.e., information about a product computed asynchronously by the data quality analysis.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get productstatuses (response)

Fields

google_expiration_date

Date on which the item expires in Google Shopping, in ISO 8601 format.

destination_statuses

The intended destinations for the product.

title

The title of the product.

kind

Identifies what kind of resource this is. Value: the fixed string "content#productStatus".

last_update_date

Date on which the item has been last updated, in ISO 8601 format.

link

The link to the product.

creation_date

Date on which the item has been created, in ISO 8601 format.

data_quality_issues

A list of data quality issues associated with the product.

product_id

The id of the product for which status is reported.