Struct google_content2::RateGroup

pub struct RateGroup {
    pub applicable_shipping_labels: Option<Vec<String>>,
    pub carrier_rates: Option<Vec<CarrierRate>>,
    pub single_value: Option<Value>,
    pub subtables: Option<Vec<Table>>,
    pub main_table: Option<Table>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
applicable_shipping_labels: Option<Vec<String>>

A list of shipping labels defining the products to which this rate group applies to. This is a disjunction: only one of the labels has to match for the rate group to apply. May only be empty for the last rate group of a service. Required.

carrier_rates: Option<Vec<CarrierRate>>

A list of carrier rates that can be referred to by mainTable or singleValue.

single_value: Option<Value>

The value of the rate group (e.g. flat rate $10). Can only be set if mainTable and subtables are not set.

subtables: Option<Vec<Table>>

A list of subtables referred to by mainTable. Can only be set if mainTable is set.

main_table: Option<Table>

A table defining the rate group, when singleValue is not expressive enough. Can only be set if singleValue is not set.