Struct google_content2::TestOrderLineItemProduct

pub struct TestOrderLineItemProduct {
    pub target_country: Option<i64>,
    pub content_language: Option<String>,
    pub title: Option<String>,
    pub mpn: Option<String>,
    pub variant_attributes: Option<Vec<OrderLineItemProductVariantAttribute>>,
    pub brand: Option<String>,
    pub image_link: Option<String>,
    pub offer_id: Option<String>,
    pub item_group_id: Option<String>,
    pub gtin: Option<String>,
    pub condition: Option<String>,
    pub price: Option<Price>,
    pub channel: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
target_country: Option<i64>

The CLDR territory code of the target country of the product.

content_language: Option<String>

The two-letter ISO 639-1 language code for the item.

title: Option<String>

The title of the product.

mpn: Option<String>

Manufacturer Part Number (MPN) of the item. Optional.

variant_attributes: Option<Vec<OrderLineItemProductVariantAttribute>>

Variant attributes for the item. Optional.

brand: Option<String>

Brand of the item.

image_link: Option<String>

URL of an image of the item.

offer_id: Option<String>

An identifier of the item.

item_group_id: Option<String>

Shared identifier for all variants of the same product. Optional.

gtin: Option<String>

Global Trade Item Number (GTIN) of the item. Optional.

condition: Option<String>

Condition or state of the item.

price: Option<Price>

The price for the product.

channel: Option<String>

The item's channel.