Struct google_content2::Value

pub struct Value {
    pub carrier_rate_name: Option<String>,
    pub no_shipping: Option<bool>,
    pub price_percentage: Option<String>,
    pub flat_rate: Option<Price>,
    pub subtable_name: Option<String>,
}

The single value of a rate group or the value of a rate group table's cell. Exactly one of noShipping, flatRate, pricePercentage, carrierRateName, subtableName must be set.

This type is not used in any activity, and only used as part of another schema.

Fields
carrier_rate_name: Option<String>

The name of a carrier rate referring to a carrier rate defined in the same rate group. Can only be set if all other fields are not set.

no_shipping: Option<bool>

If true, then the product can't ship. Must be true when set, can only be set if all other fields are not set.

price_percentage: Option<String>

A percentage of the price represented as a number in decimal notation (e.g., "5.4"). Can only be set if all other fields are not set.

flat_rate: Option<Price>

A flat rate. Can only be set if all other fields are not set.

subtable_name: Option<String>

The name of a subtable. Can only be set in table cells (i.e., not for single values), and only if all other fields are not set.