Crate google_content2

This documentation was generated from Shopping Content crate version 0.1.8+20150421, where 20150421 is the exact revision of the content:v2 schema built by the mako code generator v0.1.8.

Everything else about the Shopping Content v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

authinfo, custombatch, delete, get, insert, list, patch and update

accountshipping

custombatch, get, list, patch and update

accountstatuses

custombatch, get and list

accounttax

custombatch, get, list, patch and update

datafeeds

custombatch, delete, get, insert, list, patch and update

datafeedstatuses

custombatch, get and list

inventory

custombatch and set

products

custombatch, delete, get, insert and list

productstatuses

custombatch, get and list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.accounts().custombatch(...).doit()
let r = hub.accounts().patch(...).doit()
let r = hub.accounts().get(...).doit()
let r = hub.accounts().update(...).doit()
let r = hub.accounts().list(...).doit()
let r = hub.accounts().authinfo(...).doit()
let r = hub.accounts().delete(...).doit()
let r = hub.accounts().insert(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-content2 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_content2 as content2;
use content2::Account;
use content2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2::ShoppingContent;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Account::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().patch(req, "merchantId", "accountId")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Account
Account data.

AccountAdwordsLink
There is no detailed description.

AccountAuthinfoCall
Returns information about the authenticated user.

AccountCustombatchCall
Retrieves, inserts, updates, and deletes multiple Merchant Center (sub-)accounts in a single request.

AccountDeleteCall
Deletes a Merchant Center sub-account.

AccountGetCall
Retrieves a Merchant Center account.

AccountIdentifier
There is no detailed description.

AccountInsertCall
Creates a Merchant Center sub-account.

AccountListCall
Lists the sub-accounts in your Merchant Center account.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the ShoppingContent hub.

AccountPatchCall
Updates a Merchant Center account. This method supports patch semantics.

AccountShipping
The shipping settings of a merchant account.

AccountShippingCarrierRate
A carrier-calculated shipping rate.

AccountShippingCondition
There is no detailed description.

AccountShippingLocationGroup
A user-defined locations group in a given country. All the locations of the group must be of the same type.

AccountShippingPostalCodeRange
A postal code range, that can be either: - A range of postal codes (e.g., start=12340, end=12359) - A range of postal codes prefixes (e.g., start=1234* end=1235*). Prefixes must be of the same length (e.g., start=12* end=2* is invalid).

AccountShippingRateTable
A single or bi-dimensional table of shipping rates. Each dimension is defined in terms of consecutive price/weight ranges, delivery locations, or shipping labels.

AccountShippingRateTableCell
There is no detailed description.

AccountShippingShippingService
Shipping services provided in a country.

AccountShippingShippingServiceCalculationMethod
Shipping cost calculation method. Exactly one of the field is set.

AccountShippingShippingServiceCostRule
Building block of the cost calculation decision tree. - The tree root should have no condition and no calculation method. - All the children must have a condition on the same dimension. The first child matching a condition is entered, therefore, price and weight conditions form contiguous intervals. - The last child of an element must have no condition and matches all elements not previously matched. - Children and calculation method are mutually exclusive, and exactly one of them must be defined; the root must only have children.

AccountStatus
The status of an account, i.e., information about its products, which is computed offline and not returned immediately at insertion time.

AccountStatusDataQualityIssue
There is no detailed description.

AccountStatusExampleItem
An example of an item that has poor data quality. An item value on the landing page differs from what is submitted, or conflicts with a policy.

AccountTax
The tax settings of a merchant account.

AccountTaxTaxRule
Tax calculation rule to apply in a state or province (USA only).

AccountUpdateCall
Updates a Merchant Center account.

AccountUser
There is no detailed description.

AccountsAuthInfoResponse
There is no detailed description.

AccountsCustomBatchRequest
There is no detailed description.

AccountsCustomBatchRequestEntry
A batch entry encoding a single non-batch accounts request.

AccountsCustomBatchResponse
There is no detailed description.

AccountsCustomBatchResponseEntry
A batch entry encoding a single non-batch accounts response.

AccountsListResponse
There is no detailed description.

AccountshippingCustomBatchRequest
There is no detailed description.

AccountshippingCustomBatchRequestEntry
A batch entry encoding a single non-batch accountshipping request.

AccountshippingCustomBatchResponse
There is no detailed description.

AccountshippingCustomBatchResponseEntry
A batch entry encoding a single non-batch accountshipping response.

AccountshippingCustombatchCall
Retrieves and updates the shipping settings of multiple accounts in a single request.

AccountshippingGetCall
Retrieves the shipping settings of the account.

AccountshippingListCall
Lists the shipping settings of the sub-accounts in your Merchant Center account.

AccountshippingListResponse
There is no detailed description.

AccountshippingMethods
A builder providing access to all methods supported on accountshipping resources. It is not used directly, but through the ShoppingContent hub.

AccountshippingPatchCall
Updates the shipping settings of the account. This method supports patch semantics.

AccountshippingUpdateCall
Updates the shipping settings of the account.

AccountstatuseCustombatchCall
A builder for the custombatch method supported by a accountstatuse resource. It is not used directly, but through a AccountstatuseMethods instance.

AccountstatuseGetCall
Retrieves the status of a Merchant Center account.

AccountstatuseListCall
Lists the statuses of the sub-accounts in your Merchant Center account.

AccountstatuseMethods
A builder providing access to all methods supported on accountstatuse resources. It is not used directly, but through the ShoppingContent hub.

AccountstatusesCustomBatchRequest
There is no detailed description.

AccountstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch accountstatuses request.

AccountstatusesCustomBatchResponse
There is no detailed description.

AccountstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch accountstatuses response.

AccountstatusesListResponse
There is no detailed description.

AccounttaxCustomBatchRequest
There is no detailed description.

AccounttaxCustomBatchRequestEntry
A batch entry encoding a single non-batch accounttax request.

AccounttaxCustomBatchResponse
There is no detailed description.

AccounttaxCustomBatchResponseEntry
A batch entry encoding a single non-batch accounttax response.

AccounttaxCustombatchCall
Retrieves and updates tax settings of multiple accounts in a single request.

AccounttaxGetCall
Retrieves the tax settings of the account.

AccounttaxListCall
Lists the tax settings of the sub-accounts in your Merchant Center account.

AccounttaxListResponse
There is no detailed description.

AccounttaxMethods
A builder providing access to all methods supported on accounttax resources. It is not used directly, but through the ShoppingContent hub.

AccounttaxPatchCall
Updates the tax settings of the account. This method supports patch semantics.

AccounttaxUpdateCall
Updates the tax settings of the account.

Datafeed
Datafeed data.

DatafeedCustombatchCall
A builder for the custombatch method supported by a datafeed resource. It is not used directly, but through a DatafeedMethods instance.

DatafeedDeleteCall
Deletes a datafeed from your Merchant Center account.

DatafeedFetchSchedule
The required fields vary based on the frequency of fetching. For a monthly fetch schedule, day_of_month and hour are required. For a weekly fetch schedule, weekday and hour are required. For a daily fetch schedule, only hour is required.

DatafeedFormat
There is no detailed description.

DatafeedGetCall
Retrieves a datafeed from your Merchant Center account.

DatafeedInsertCall
Registers a datafeed with your Merchant Center account.

DatafeedListCall
Lists the datafeeds in your Merchant Center account.

DatafeedMethods
A builder providing access to all methods supported on datafeed resources. It is not used directly, but through the ShoppingContent hub.

DatafeedPatchCall
Updates a datafeed of your Merchant Center account. This method supports patch semantics.

DatafeedStatus
The status of a datafeed, i.e., the result of the last retrieval of the datafeed computed asynchronously when the feed processing is finished.

DatafeedStatusError
An error occurring in the feed, like "invalid price".

DatafeedStatusExample
An example occurrence for a particular error.

DatafeedUpdateCall
Updates a datafeed of your Merchant Center account.

DatafeedsCustomBatchRequest
There is no detailed description.

DatafeedsCustomBatchRequestEntry
A batch entry encoding a single non-batch datafeeds request.

DatafeedsCustomBatchResponse
There is no detailed description.

DatafeedsCustomBatchResponseEntry
A batch entry encoding a single non-batch datafeeds response.

DatafeedsListResponse
There is no detailed description.

DatafeedstatuseCustombatchCall
A builder for the custombatch method supported by a datafeedstatuse resource. It is not used directly, but through a DatafeedstatuseMethods instance.

DatafeedstatuseGetCall
Retrieves the status of a datafeed from your Merchant Center account.

DatafeedstatuseListCall
Lists the statuses of the datafeeds in your Merchant Center account.

DatafeedstatuseMethods
A builder providing access to all methods supported on datafeedstatuse resources. It is not used directly, but through the ShoppingContent hub.

DatafeedstatusesCustomBatchRequest
There is no detailed description.

DatafeedstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch datafeedstatuses request.

DatafeedstatusesCustomBatchResponse
There is no detailed description.

DatafeedstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch datafeedstatuses response.

DatafeedstatusesListResponse
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ErrorType
An error returned by the API.

Errors
A list of errors returned by a failed batch entry.

Inventory
There is no detailed description.

InventoryCustomBatchRequest
There is no detailed description.

InventoryCustomBatchRequestEntry
A batch entry encoding a single non-batch inventory request.

InventoryCustomBatchResponse
There is no detailed description.

InventoryCustomBatchResponseEntry
A batch entry encoding a single non-batch inventory response.

InventoryCustombatchCall
Updates price and availability for multiple products or stores in a single request.

InventoryMethods
A builder providing access to all methods supported on inventory resources. It is not used directly, but through the ShoppingContent hub.

InventorySetCall
Updates price and availability of a product in your Merchant Center account.

InventorySetRequest
There is no detailed description.

InventorySetResponse
There is no detailed description.

LoyaltyPoints
There is no detailed description.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Price
There is no detailed description.

Product
Product data.

ProductAspect
There is no detailed description.

ProductCustomAttribute
There is no detailed description.

ProductCustomGroup
There is no detailed description.

ProductCustombatchCall
Retrieves, inserts, and deletes multiple products in a single request.

ProductDeleteCall
Deletes a product from your Merchant Center account.

ProductDestination
There is no detailed description.

ProductGetCall
Retrieves a product from your Merchant Center account.

ProductInsertCall
Uploads a product to your Merchant Center account.

ProductInstallment
There is no detailed description.

ProductListCall
Lists the products in your Merchant Center account.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the ShoppingContent hub.

ProductShipping
There is no detailed description.

ProductShippingDimension
There is no detailed description.

ProductShippingWeight
There is no detailed description.

ProductStatus
The status of a product, i.e., information about a product computed asynchronously by the data quality analysis.

ProductStatusDataQualityIssue
There is no detailed description.

ProductStatusDestinationStatus
There is no detailed description.

ProductTax
There is no detailed description.

ProductUnitPricingBaseMeasure
There is no detailed description.

ProductUnitPricingMeasure
There is no detailed description.

ProductsCustomBatchRequest
There is no detailed description.

ProductsCustomBatchRequestEntry
A batch entry encoding a single non-batch products request.

ProductsCustomBatchResponse
There is no detailed description.

ProductsCustomBatchResponseEntry
A batch entry encoding a single non-batch products response.

ProductsListResponse
There is no detailed description.

ProductstatuseCustombatchCall
Gets the statuses of multiple products in a single request.

ProductstatuseGetCall
Gets the status of a product from your Merchant Center account.

ProductstatuseListCall
Lists the statuses of the products in your Merchant Center account.

ProductstatuseMethods
A builder providing access to all methods supported on productstatuse resources. It is not used directly, but through the ShoppingContent hub.

ProductstatusesCustomBatchRequest
There is no detailed description.

ProductstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch productstatuses request.

ProductstatusesCustomBatchResponse
There is no detailed description.

ProductstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch productstatuses response.

ProductstatusesListResponse
There is no detailed description.

ShoppingContent
Central instance to access all ShoppingContent related resource activities

Weight
There is no detailed description.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.