Crate google_content2

This documentation was generated from Shopping Content crate version 1.0.9+20190702, where 20190702 is the exact revision of the content:v2 schema built by the mako code generator v1.0.9.

Everything else about the Shopping Content v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

accounts

authinfo, claimwebsite, custombatch, delete, get, insert, link, list, patch and update

accountstatuses

custombatch, get and list

accounttax

custombatch, get, list, patch and update

datafeeds

custombatch, delete, fetchnow, get, insert, list, patch and update

datafeedstatuses

custombatch, get and list

inventory

custombatch and set

liasettings

custombatch, get, getaccessiblegmbaccounts, list, listposdataproviders, patch, requestgmbaccess, requestinventoryverification, setinventoryverificationcontact, setposdataprovider and update

orderinvoices

createchargeinvoice and createrefundinvoice

orderpayments

notifyauthapproved, notifyauthdeclined, notifycharge and notifyrefund

orderreports

listdisbursements and listtransactions

orderreturns

get and list

orders

acknowledge, advancetestorder, cancel, cancellineitem, canceltestorderbycustomer, createtestorder, createtestreturn, custombatch, get, getbymerchantorderid, gettestordertemplate, instorerefundlineitem, list, refund, rejectreturnlineitem, returnlineitem, returnrefundlineitem, setlineitemmetadata, shiplineitems, updatelineitemshippingdetails, updatemerchantorderid and updateshipment

pos

custombatch, delete, get, insert, inventory, list and sale

products

custombatch, delete, get, insert and list

productstatuses

custombatch, get and list

shippingsettings

custombatch, get, getsupportedcarriers, getsupportedholidays, list, patch and update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.orders().list(...).doit()
let r = hub.orders().updatemerchantorderid(...).doit()
let r = hub.orders().gettestordertemplate(...).doit()
let r = hub.orders().updatelineitemshippingdetails(...).doit()
let r = hub.orders().refund(...).doit()
let r = hub.orders().acknowledge(...).doit()
let r = hub.orders().instorerefundlineitem(...).doit()
let r = hub.orders().advancetestorder(...).doit()
let r = hub.orders().rejectreturnlineitem(...).doit()
let r = hub.orders().createtestorder(...).doit()
let r = hub.orders().cancellineitem(...).doit()
let r = hub.orders().canceltestorderbycustomer(...).doit()
let r = hub.orders().returnrefundlineitem(...).doit()
let r = hub.orders().get(...).doit()
let r = hub.orders().returnlineitem(...).doit()
let r = hub.orders().custombatch(...).doit()
let r = hub.orders().getbymerchantorderid(...).doit()
let r = hub.orders().shiplineitems(...).doit()
let r = hub.orders().createtestreturn(...).doit()
let r = hub.orders().updateshipment(...).doit()
let r = hub.orders().setlineitemmetadata(...).doit()
let r = hub.orders().cancel(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-content2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_content2 as content2;
use content2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use content2::ShoppingContent;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ShoppingContent::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.orders().list("merchantId")
             .add_statuses("justo")
             .placed_date_start("amet.")
             .placed_date_end("erat")
             .page_token("labore")
             .order_by("sea")
             .max_results(11)
             .acknowledged(true)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Account
Account data. After the creation of a new account it may take a few minutes before it is fully operational. The methods delete, insert, patch, and update require the admin role.

AccountAddress
There is no detailed description.

AccountAdwordsLink
There is no detailed description.

AccountAuthinfoCall
Returns information about the authenticated user.

AccountBusinessInformation
There is no detailed description.

AccountClaimwebsiteCall
Claims the website of a Merchant Center sub-account.

AccountCustombatchCall
Retrieves, inserts, updates, and deletes multiple Merchant Center (sub-)accounts in a single request.

AccountCustomerService
There is no detailed description.

AccountDeleteCall
Deletes a Merchant Center sub-account.

AccountGetCall
Retrieves a Merchant Center account.

AccountGoogleMyBusinessLink
There is no detailed description.

AccountIdentifier
There is no detailed description.

AccountInsertCall
Creates a Merchant Center sub-account.

AccountLinkCall
Performs an action on a link between two Merchant Center accounts, namely accountId and linkedAccountId.

AccountListCall
Lists the sub-accounts in your Merchant Center account.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the ShoppingContent hub.

AccountPatchCall
Updates a Merchant Center account. This method supports patch semantics.

AccountStatus
The status of an account, i.e., information about its products, which is computed offline and not returned immediately at insertion time.

AccountStatusAccountLevelIssue
There is no detailed description.

AccountStatusDataQualityIssue
There is no detailed description.

AccountStatusExampleItem
There is no detailed description.

AccountStatusItemLevelIssue
There is no detailed description.

AccountStatusProducts
There is no detailed description.

AccountStatusStatistics
There is no detailed description.

AccountTax
The tax settings of a merchant account. All methods require the admin role.

AccountTaxTaxRule
Tax calculation rule to apply in a state or province (USA only).

AccountUpdateCall
Updates a Merchant Center account.

AccountUser
There is no detailed description.

AccountYouTubeChannelLink
There is no detailed description.

AccountsAuthInfoResponse
There is no detailed description.

AccountsClaimWebsiteResponse
There is no detailed description.

AccountsCustomBatchRequest
There is no detailed description.

AccountsCustomBatchRequestEntry
A batch entry encoding a single non-batch accounts request.

AccountsCustomBatchRequestEntryLinkRequest
There is no detailed description.

AccountsCustomBatchResponse
There is no detailed description.

AccountsCustomBatchResponseEntry
A batch entry encoding a single non-batch accounts response.

AccountsLinkRequest
There is no detailed description.

AccountsLinkResponse
There is no detailed description.

AccountsListResponse
There is no detailed description.

AccountstatuseCustombatchCall
Retrieves multiple Merchant Center account statuses in a single request.

AccountstatuseGetCall
Retrieves the status of a Merchant Center account. No itemLevelIssues are returned for multi-client accounts.

AccountstatuseListCall
Lists the statuses of the sub-accounts in your Merchant Center account.

AccountstatuseMethods
A builder providing access to all methods supported on accountstatuse resources. It is not used directly, but through the ShoppingContent hub.

AccountstatusesCustomBatchRequest
There is no detailed description.

AccountstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch accountstatuses request.

AccountstatusesCustomBatchResponse
There is no detailed description.

AccountstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch accountstatuses response.

AccountstatusesListResponse
There is no detailed description.

AccounttaxCustomBatchRequest
There is no detailed description.

AccounttaxCustomBatchRequestEntry
A batch entry encoding a single non-batch accounttax request.

AccounttaxCustomBatchResponse
There is no detailed description.

AccounttaxCustomBatchResponseEntry
A batch entry encoding a single non-batch accounttax response.

AccounttaxCustombatchCall
Retrieves and updates tax settings of multiple accounts in a single request.

AccounttaxGetCall
Retrieves the tax settings of the account.

AccounttaxListCall
Lists the tax settings of the sub-accounts in your Merchant Center account.

AccounttaxListResponse
There is no detailed description.

AccounttaxMethods
A builder providing access to all methods supported on accounttax resources. It is not used directly, but through the ShoppingContent hub.

AccounttaxPatchCall
Updates the tax settings of the account. This method supports patch semantics.

AccounttaxUpdateCall
Updates the tax settings of the account.

Amount
There is no detailed description.

BusinessDayConfig
There is no detailed description.

CarrierRate
There is no detailed description.

CarriersCarrier
There is no detailed description.

Chunk

ContentRange
Implements the Content-Range header, for serialization only

CustomAttribute
There is no detailed description.

CustomGroup
There is no detailed description.

CustomerReturnReason
There is no detailed description.

CutoffTime
There is no detailed description.

Datafeed
Datafeed configuration data.

DatafeedCustombatchCall
Deletes, fetches, gets, inserts and updates multiple datafeeds in a single request.

DatafeedDeleteCall
Deletes a datafeed configuration from your Merchant Center account.

DatafeedFetchSchedule
The required fields vary based on the frequency of fetching. For a monthly fetch schedule, day_of_month and hour are required. For a weekly fetch schedule, weekday and hour are required. For a daily fetch schedule, only hour is required.

DatafeedFetchnowCall
Invokes a fetch for the datafeed in your Merchant Center account.

DatafeedFormat
There is no detailed description.

DatafeedGetCall
Retrieves a datafeed configuration from your Merchant Center account.

DatafeedInsertCall
Registers a datafeed configuration with your Merchant Center account.

DatafeedListCall
Lists the configurations for datafeeds in your Merchant Center account.

DatafeedMethods
A builder providing access to all methods supported on datafeed resources. It is not used directly, but through the ShoppingContent hub.

DatafeedPatchCall
Updates a datafeed configuration of your Merchant Center account. This method supports patch semantics.

DatafeedStatus
The status of a datafeed, i.e., the result of the last retrieval of the datafeed computed asynchronously when the feed processing is finished.

DatafeedStatusError
An error occurring in the feed, like "invalid price".

DatafeedStatusExample
An example occurrence for a particular error.

DatafeedTarget
There is no detailed description.

DatafeedUpdateCall
Updates a datafeed configuration of your Merchant Center account.

DatafeedsCustomBatchRequest
There is no detailed description.

DatafeedsCustomBatchRequestEntry
A batch entry encoding a single non-batch datafeeds request.

DatafeedsCustomBatchResponse
There is no detailed description.

DatafeedsCustomBatchResponseEntry
A batch entry encoding a single non-batch datafeeds response.

DatafeedsFetchNowResponse
There is no detailed description.

DatafeedsListResponse
There is no detailed description.

DatafeedstatuseCustombatchCall
Gets multiple Merchant Center datafeed statuses in a single request.

DatafeedstatuseGetCall
Retrieves the status of a datafeed from your Merchant Center account.

DatafeedstatuseListCall
Lists the statuses of the datafeeds in your Merchant Center account.

DatafeedstatuseMethods
A builder providing access to all methods supported on datafeedstatuse resources. It is not used directly, but through the ShoppingContent hub.

DatafeedstatusesCustomBatchRequest
There is no detailed description.

DatafeedstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch datafeedstatuses request.

DatafeedstatusesCustomBatchResponse
There is no detailed description.

DatafeedstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch datafeedstatuses response.

DatafeedstatusesListResponse
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeliveryTime
There is no detailed description.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ErrorType
An error returned by the API.

Errors
A list of errors returned by a failed batch entry.

GmbAccounts
There is no detailed description.

GmbAccountsGmbAccount
There is no detailed description.

Headers
A non-empty list of row or column headers for a table. Exactly one of prices, weights, numItems, postalCodeGroupNames, or location must be set.

HolidayCutoff
There is no detailed description.

HolidaysHoliday
There is no detailed description.

Installment
There is no detailed description.

Inventory
There is no detailed description.

InventoryCustomBatchRequest
There is no detailed description.

InventoryCustomBatchRequestEntry
A batch entry encoding a single non-batch inventory request.

InventoryCustomBatchResponse
There is no detailed description.

InventoryCustomBatchResponseEntry
A batch entry encoding a single non-batch inventory response.

InventoryCustombatchCall
Updates price and availability for multiple products or stores in a single request. This operation does not update the expiration date of the products.

InventoryMethods
A builder providing access to all methods supported on inventory resources. It is not used directly, but through the ShoppingContent hub.

InventoryPickup
There is no detailed description.

InventorySetCall
Updates price and availability of a product in your Merchant Center account.

InventorySetRequest
There is no detailed description.

InventorySetResponse
There is no detailed description.

InvoiceSummary
There is no detailed description.

InvoiceSummaryAdditionalChargeSummary
There is no detailed description.

JsonServerError
A utility type which can decode a server response that indicates error

LiaAboutPageSettings
There is no detailed description.

LiaCountrySettings
There is no detailed description.

LiaInventorySettings
There is no detailed description.

LiaOnDisplayToOrderSettings
There is no detailed description.

LiaPosDataProvider
There is no detailed description.

LiaSettings
Local Inventory ads (LIA) settings. All methods except listposdataproviders require the admin role.

LiasettingCustombatchCall
Retrieves and/or updates the LIA settings of multiple accounts in a single request.

LiasettingGetCall
Retrieves the LIA settings of the account.

LiasettingGetaccessiblegmbaccountCall
Retrieves the list of accessible Google My Business accounts.

LiasettingListCall
Lists the LIA settings of the sub-accounts in your Merchant Center account.

LiasettingListposdataproviderCall
Retrieves the list of POS data providers that have active settings for the all eiligible countries.

LiasettingMethods
A builder providing access to all methods supported on liasetting resources. It is not used directly, but through the ShoppingContent hub.

LiasettingPatchCall
Updates the LIA settings of the account. This method supports patch semantics.

LiasettingRequestgmbaccesCall
Requests access to a specified Google My Business account.

LiasettingRequestinventoryverificationCall
Requests inventory validation for the specified country.

LiasettingSetinventoryverificationcontactCall
Sets the inventory verification contract for the specified country.

LiasettingSetposdataproviderCall
Sets the POS data provider for the specified country.

LiasettingUpdateCall
Updates the LIA settings of the account.

LiasettingsCustomBatchRequest
There is no detailed description.

LiasettingsCustomBatchRequestEntry
There is no detailed description.

LiasettingsCustomBatchResponse
There is no detailed description.

LiasettingsCustomBatchResponseEntry
There is no detailed description.

LiasettingsGetAccessibleGmbAccountsResponse
There is no detailed description.

LiasettingsListPosDataProvidersResponse
There is no detailed description.

LiasettingsListResponse
There is no detailed description.

LiasettingsRequestGmbAccessResponse
There is no detailed description.

LiasettingsRequestInventoryVerificationResponse
There is no detailed description.

LiasettingsSetInventoryVerificationContactResponse
There is no detailed description.

LiasettingsSetPosDataProviderResponse
There is no detailed description.

LocationIdSet
There is no detailed description.

LoyaltyPoints
There is no detailed description.

MerchantOrderReturn
There is no detailed description.

MerchantOrderReturnItem
There is no detailed description.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Order
Order. All methods require the order manager role.

OrderAcknowledgeCall
Marks an order as acknowledged.

OrderAddress
There is no detailed description.

OrderAdvancetestorderCall
Sandbox only. Moves a test order from state "inProgress" to state "pendingShipment".

OrderCancelCall
Cancels all line items in an order, making a full refund.

OrderCancellation
There is no detailed description.

OrderCancellineitemCall
Cancels a line item, making a full refund.

OrderCanceltestorderbycustomerCall
Sandbox only. Cancels a test order for customer-initiated cancellation.

OrderCreatetestorderCall
Sandbox only. Creates a test order.

OrderCreatetestreturnCall
Sandbox only. Creates a test return.

OrderCustombatchCall
Retrieves or modifies multiple orders in a single request.

OrderCustomer
There is no detailed description.

OrderCustomerMarketingRightsInfo
There is no detailed description.

OrderDeliveryDetails
There is no detailed description.

OrderGetCall
Retrieves an order from your Merchant Center account.

OrderGetbymerchantorderidCall
Retrieves an order using merchant order ID.

OrderGettestordertemplateCall
Sandbox only. Retrieves an order template that can be used to quickly create a new order in sandbox.

OrderInstorerefundlineitemCall
Notifies that item return and refund was handled directly by merchant outside of Google payments processing (e.g. cash refund done in store). Note: We recommend calling the returnrefundlineitem method to refund in-store returns. We will issue the refund directly to the customer. This helps to prevent possible differences arising between merchant and Google transaction records. We also recommend having the point of sale system communicate with Google to ensure that customers do not receive a double refund by first refunding via Google then via an in-store return.

OrderLegacyPromotion
There is no detailed description.

OrderLegacyPromotionBenefit
There is no detailed description.

OrderLineItem
There is no detailed description.

OrderLineItemProduct
There is no detailed description.

OrderLineItemProductFee
There is no detailed description.

OrderLineItemProductVariantAttribute
There is no detailed description.

OrderLineItemReturnInfo
There is no detailed description.

OrderLineItemShippingDetails
There is no detailed description.

OrderLineItemShippingDetailsMethod
There is no detailed description.

OrderListCall
Lists the orders in your Merchant Center account.

OrderMerchantProvidedAnnotation
There is no detailed description.

OrderMethods
A builder providing access to all methods supported on order resources. It is not used directly, but through the ShoppingContent hub.

OrderPaymentMethod
There is no detailed description.

OrderRefund
There is no detailed description.

OrderRefundCall
Deprecated, please use returnRefundLineItem instead.

OrderRejectreturnlineitemCall
Rejects return on an line item.

OrderReportDisbursement
Order disbursement. All methods require the payment analyst role.

OrderReportTransaction
There is no detailed description.

OrderReturn
There is no detailed description.

OrderReturnlineitemCall
Returns a line item.

OrderReturnrefundlineitemCall
Returns and refunds a line item. Note that this method can only be called on fully shipped orders.

OrderSetlineitemmetadataCall
Sets (or overrides if it already exists) merchant provided annotations in the form of key-value pairs. A common use case would be to supply us with additional structured information about a line item that cannot be provided via other methods. Submitted key-value pairs can be retrieved as part of the orders resource.

OrderShiplineitemCall
Marks line item(s) as shipped.

OrderShipment
There is no detailed description.

OrderShipmentLineItemShipment
There is no detailed description.

OrderUpdatelineitemshippingdetailCall
Updates ship by and delivery by dates for a line item.

OrderUpdatemerchantorderidCall
Updates the merchant order ID for a given order.

OrderUpdateshipmentCall
Updates a shipment's status, carrier, and/or tracking ID.

OrderinvoiceCreatechargeinvoiceCall
Creates a charge invoice for a shipment group, and triggers a charge capture for orderinvoice enabled orders.

OrderinvoiceCreaterefundinvoiceCall
Creates a refund invoice for one or more shipment groups, and triggers a refund for orderinvoice enabled orders. This can only be used for line items that have previously been charged using createChargeInvoice. All amounts (except for the summary) are incremental with respect to the previous invoice.

OrderinvoiceMethods
A builder providing access to all methods supported on orderinvoice resources. It is not used directly, but through the ShoppingContent hub.

OrderinvoicesCreateChargeInvoiceRequest
There is no detailed description.

OrderinvoicesCreateChargeInvoiceResponse
There is no detailed description.

OrderinvoicesCreateRefundInvoiceRequest
There is no detailed description.

OrderinvoicesCreateRefundInvoiceResponse
There is no detailed description.

OrderinvoicesCustomBatchRequestEntryCreateRefundInvoiceRefundOption
There is no detailed description.

OrderinvoicesCustomBatchRequestEntryCreateRefundInvoiceReturnOption
There is no detailed description.

OrderpaymentMethods
A builder providing access to all methods supported on orderpayment resources. It is not used directly, but through the ShoppingContent hub.

OrderpaymentNotifyauthapprovedCall
Notify about successfully authorizing user's payment method for a given amount.

OrderpaymentNotifyauthdeclinedCall
Notify about failure to authorize user's payment method.

OrderpaymentNotifychargeCall
Notify about charge on user's selected payments method.

OrderpaymentNotifyrefundCall
Notify about refund on user's selected payments method.

OrderpaymentsNotifyAuthApprovedRequest
There is no detailed description.

OrderpaymentsNotifyAuthApprovedResponse
There is no detailed description.

OrderpaymentsNotifyAuthDeclinedRequest
There is no detailed description.

OrderpaymentsNotifyAuthDeclinedResponse
There is no detailed description.

OrderpaymentsNotifyChargeRequest
There is no detailed description.

OrderpaymentsNotifyChargeResponse
There is no detailed description.

OrderpaymentsNotifyRefundRequest
There is no detailed description.

OrderpaymentsNotifyRefundResponse
There is no detailed description.

OrderreportListdisbursementCall
Retrieves a report for disbursements from your Merchant Center account.

OrderreportListtransactionCall
Retrieves a list of transactions for a disbursement from your Merchant Center account.

OrderreportMethods
A builder providing access to all methods supported on orderreport resources. It is not used directly, but through the ShoppingContent hub.

OrderreportsListDisbursementsResponse
There is no detailed description.

OrderreportsListTransactionsResponse
There is no detailed description.

OrderreturnGetCall
Retrieves an order return from your Merchant Center account.

OrderreturnListCall
Lists order returns in your Merchant Center account.

OrderreturnMethods
A builder providing access to all methods supported on orderreturn resources. It is not used directly, but through the ShoppingContent hub.

OrderreturnsListResponse
There is no detailed description.

OrdersAcknowledgeRequest
There is no detailed description.

OrdersAcknowledgeResponse
There is no detailed description.

OrdersAdvanceTestOrderResponse
There is no detailed description.

OrdersCancelLineItemRequest
There is no detailed description.

OrdersCancelLineItemResponse
There is no detailed description.

OrdersCancelRequest
There is no detailed description.

OrdersCancelResponse
There is no detailed description.

OrdersCancelTestOrderByCustomerRequest
There is no detailed description.

OrdersCancelTestOrderByCustomerResponse
There is no detailed description.

OrdersCreateTestOrderRequest
There is no detailed description.

OrdersCreateTestOrderResponse
There is no detailed description.

OrdersCreateTestReturnRequest
There is no detailed description.

OrdersCreateTestReturnResponse
There is no detailed description.

OrdersCustomBatchRequest
There is no detailed description.

OrdersCustomBatchRequestEntry
There is no detailed description.

OrdersCustomBatchRequestEntryCancel
There is no detailed description.

OrdersCustomBatchRequestEntryCancelLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryCreateTestReturnReturnItem
There is no detailed description.

OrdersCustomBatchRequestEntryInStoreRefundLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryRefund
There is no detailed description.

OrdersCustomBatchRequestEntryRejectReturnLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryReturnLineItem
There is no detailed description.

OrdersCustomBatchRequestEntryReturnRefundLineItem
There is no detailed description.

OrdersCustomBatchRequestEntrySetLineItemMetadata
There is no detailed description.

OrdersCustomBatchRequestEntryShipLineItems
There is no detailed description.

OrdersCustomBatchRequestEntryShipLineItemsShipmentInfo
There is no detailed description.

OrdersCustomBatchRequestEntryUpdateLineItemShippingDetails
There is no detailed description.

OrdersCustomBatchRequestEntryUpdateShipment
There is no detailed description.

OrdersCustomBatchResponse
There is no detailed description.

OrdersCustomBatchResponseEntry
There is no detailed description.

OrdersGetByMerchantOrderIdResponse
There is no detailed description.

OrdersGetTestOrderTemplateResponse
There is no detailed description.

OrdersInStoreRefundLineItemRequest
There is no detailed description.

OrdersInStoreRefundLineItemResponse
There is no detailed description.

OrdersListResponse
There is no detailed description.

OrdersRefundRequest
There is no detailed description.

OrdersRefundResponse
There is no detailed description.

OrdersRejectReturnLineItemRequest
There is no detailed description.

OrdersRejectReturnLineItemResponse
There is no detailed description.

OrdersReturnLineItemRequest
There is no detailed description.

OrdersReturnLineItemResponse
There is no detailed description.

OrdersReturnRefundLineItemRequest
There is no detailed description.

OrdersReturnRefundLineItemResponse
There is no detailed description.

OrdersSetLineItemMetadataRequest
There is no detailed description.

OrdersSetLineItemMetadataResponse
There is no detailed description.

OrdersShipLineItemsRequest
There is no detailed description.

OrdersShipLineItemsResponse
There is no detailed description.

OrdersUpdateLineItemShippingDetailsRequest
There is no detailed description.

OrdersUpdateLineItemShippingDetailsResponse
There is no detailed description.

OrdersUpdateMerchantOrderIdRequest
There is no detailed description.

OrdersUpdateMerchantOrderIdResponse
There is no detailed description.

OrdersUpdateShipmentRequest
There is no detailed description.

OrdersUpdateShipmentResponse
There is no detailed description.

PoCustombatchCall
Batches multiple POS-related calls in a single request.

PoDeleteCall
Deletes a store for the given merchant.

PoGetCall
Retrieves information about the given store.

PoInsertCall
Creates a store for the given merchant.

PoInventoryCall
Submit inventory for the given merchant.

PoListCall
Lists the stores of the target merchant.

PoMethods
A builder providing access to all methods supported on po resources. It is not used directly, but through the ShoppingContent hub.

PoSaleCall
Submit a sale event for the given merchant.

PosCustomBatchRequest
There is no detailed description.

PosCustomBatchRequestEntry
There is no detailed description.

PosCustomBatchResponse
There is no detailed description.

PosCustomBatchResponseEntry
There is no detailed description.

PosDataProviders
There is no detailed description.

PosDataProvidersPosDataProvider
There is no detailed description.

PosInventory
The absolute quantity of an item available at the given store.

PosInventoryRequest
There is no detailed description.

PosInventoryResponse
There is no detailed description.

PosListResponse
There is no detailed description.

PosSale
The change of the available quantity of an item at the given store.

PosSaleRequest
There is no detailed description.

PosSaleResponse
There is no detailed description.

PosStore
Store resource.

PostalCodeGroup
There is no detailed description.

PostalCodeRange
There is no detailed description.

Price
There is no detailed description.

Product
Product data. After inserting, updating, or deleting a product, it may take several minutes before changes take effect.

ProductAmount
There is no detailed description.

ProductAspect
There is no detailed description.

ProductCustombatchCall
Retrieves, inserts, and deletes multiple products in a single request.

ProductDeleteCall
Deletes a product from your Merchant Center account.

ProductDestination
There is no detailed description.

ProductGetCall
Retrieves a product from your Merchant Center account.

ProductInsertCall
Uploads a product to your Merchant Center account. If an item with the same channel, contentLanguage, offerId, and targetCountry already exists, this method updates that entry.

ProductListCall
Lists the products in your Merchant Center account.

ProductMethods
A builder providing access to all methods supported on product resources. It is not used directly, but through the ShoppingContent hub.

ProductShipping
There is no detailed description.

ProductShippingDimension
There is no detailed description.

ProductShippingWeight
There is no detailed description.

ProductStatus
The status of a product, i.e., information about a product computed asynchronously.

ProductStatusDataQualityIssue
There is no detailed description.

ProductStatusDestinationStatus
There is no detailed description.

ProductStatusItemLevelIssue
There is no detailed description.

ProductTax
There is no detailed description.

ProductUnitPricingBaseMeasure
There is no detailed description.

ProductUnitPricingMeasure
There is no detailed description.

ProductsCustomBatchRequest
There is no detailed description.

ProductsCustomBatchRequestEntry
A batch entry encoding a single non-batch products request.

ProductsCustomBatchResponse
There is no detailed description.

ProductsCustomBatchResponseEntry
A batch entry encoding a single non-batch products response.

ProductsListResponse
There is no detailed description.

ProductstatuseCustombatchCall
Gets the statuses of multiple products in a single request.

ProductstatuseGetCall
Gets the status of a product from your Merchant Center account.

ProductstatuseListCall
Lists the statuses of the products in your Merchant Center account.

ProductstatuseMethods
A builder providing access to all methods supported on productstatuse resources. It is not used directly, but through the ShoppingContent hub.

ProductstatusesCustomBatchRequest
There is no detailed description.

ProductstatusesCustomBatchRequestEntry
A batch entry encoding a single non-batch productstatuses request.

ProductstatusesCustomBatchResponse
There is no detailed description.

ProductstatusesCustomBatchResponseEntry
A batch entry encoding a single non-batch productstatuses response.

ProductstatusesListResponse
There is no detailed description.

Promotion
There is no detailed description.

RangeResponseHeader

RateGroup
There is no detailed description.

RefundReason
There is no detailed description.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ReturnShipment
There is no detailed description.

Row
There is no detailed description.

ServerError

ServerMessage

Service
There is no detailed description.

ShipmentInvoice
There is no detailed description.

ShipmentInvoiceLineItemInvoice
There is no detailed description.

ShipmentTrackingInfo
There is no detailed description.

ShippingSettings
The merchant account's shipping settings. All methods except getsupportedcarriers and getsupportedholidays require the admin role.

ShippingsettingCustombatchCall
Retrieves and updates the shipping settings of multiple accounts in a single request.

ShippingsettingGetCall
Retrieves the shipping settings of the account.

ShippingsettingGetsupportedcarrierCall
Retrieves supported carriers and carrier services for an account.

ShippingsettingGetsupportedholidayCall
Retrieves supported holidays for an account.

ShippingsettingListCall
Lists the shipping settings of the sub-accounts in your Merchant Center account.

ShippingsettingMethods
A builder providing access to all methods supported on shippingsetting resources. It is not used directly, but through the ShoppingContent hub.

ShippingsettingPatchCall
Updates the shipping settings of the account. This method supports patch semantics.

ShippingsettingUpdateCall
Updates the shipping settings of the account.

ShippingsettingsCustomBatchRequest
There is no detailed description.

ShippingsettingsCustomBatchRequestEntry
A batch entry encoding a single non-batch shippingsettings request.

ShippingsettingsCustomBatchResponse
There is no detailed description.

ShippingsettingsCustomBatchResponseEntry
A batch entry encoding a single non-batch shipping settings response.

ShippingsettingsGetSupportedCarriersResponse
There is no detailed description.

ShippingsettingsGetSupportedHolidaysResponse
There is no detailed description.

ShippingsettingsListResponse
There is no detailed description.

ShoppingContent
Central instance to access all ShoppingContent related resource activities

Table
There is no detailed description.

TestOrder
There is no detailed description.

TestOrderCustomer
There is no detailed description.

TestOrderCustomerMarketingRightsInfo
There is no detailed description.

TestOrderLineItem
There is no detailed description.

TestOrderLineItemProduct
There is no detailed description.

TestOrderPaymentMethod
There is no detailed description.

TransitTable
There is no detailed description.

TransitTableTransitTimeRow
There is no detailed description.

TransitTableTransitTimeRowTransitTimeValue
There is no detailed description.

UnitInvoice
There is no detailed description.

UnitInvoiceAdditionalCharge
There is no detailed description.

UnitInvoiceTaxLine
There is no detailed description.

Value
The single value of a rate group or the value of a rate group table's cell. Exactly one of noShipping, flatRate, pricePercentage, carrierRateName, subtableName must be set.

Weight
There is no detailed description.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.