Struct google_content2::Datafeed

pub struct Datafeed {
    pub kind: Option<String>,
    pub content_type: Option<String>,
    pub name: Option<String>,
    pub content_language: Option<String>,
    pub format: Option<DatafeedFormat>,
    pub fetch_schedule: Option<DatafeedFetchSchedule>,
    pub target_country: Option<i64>,
    pub file_name: Option<String>,
    pub id: Option<String>,
    pub intended_destinations: Option<Vec<String>>,
    pub attribute_language: Option<String>,
}

Datafeed data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert datafeeds (request|response)

custombatch datafeeds (none)

patch datafeeds (request|response)

list datafeeds (none)

delete datafeeds (none)

get datafeeds (response)

update datafeeds (request|response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "content#datafeed".

content_type

The type of data feed.

name

A descriptive name of the data feed.

content_language

The two-letter ISO 639-1 language of the items in the feed. Must be a valid language for targetCountry.

format

Format of the feed file.

fetch_schedule

Fetch schedule for the feed file.

target_country

The country where the items in the feed will be included in the search index, represented as a CLDR territory code.

file_name

The filename of the feed. All feeds must have a unique file name.

id

The ID of the data feed.

intended_destinations

The list of intended destinations (corresponds to checked check boxes in Merchant Center).

attribute_language

The two-letter ISO 639-1 language in which the attributes are defined in the data feed.