Struct google_content2::Product

pub struct Product {
    pub adult: Option<bool>,
    pub color: Option<String>,
    pub adwords_labels: Option<Vec<String>>,
    pub item_group_id: Option<String>,
    pub expiration_date: Option<String>,
    pub availability: Option<String>,
    pub adwords_grouping: Option<String>,
    pub destinations: Option<Vec<ProductDestination>>,
    pub target_country: Option<i64>,
    pub custom_label4: Option<String>,
    pub title: Option<String>,
    pub online_only: Option<bool>,
    pub mobile_link: Option<String>,
    pub aspects: Option<Vec<ProductAspect>>,
    pub shipping_height: Option<ProductShippingDimension>,
    pub content_language: Option<String>,
    pub mpn: Option<String>,
    pub sale_price_effective_date: Option<String>,
    pub brand: Option<String>,
    pub material: Option<String>,
    pub link: Option<String>,
    pub adwords_redirect: Option<String>,
    pub energy_efficiency_class: Option<String>,
    pub size_system: Option<String>,
    pub size_type: Option<String>,
    pub custom_label3: Option<String>,
    pub custom_label2: Option<String>,
    pub condition: Option<String>,
    pub custom_label0: Option<String>,
    pub kind: Option<String>,
    pub installment: Option<ProductInstallment>,
    pub sizes: Option<Vec<String>>,
    pub gender: Option<String>,
    pub taxes: Option<Vec<ProductTax>>,
    pub identifier_exists: Option<bool>,
    pub sale_price: Option<Price>,
    pub product_type: Option<String>,
    pub display_ads_similar_ids: Option<Vec<String>>,
    pub custom_label1: Option<String>,
    pub age_group: Option<String>,
    pub additional_image_links: Option<Vec<String>>,
    pub offer_id: Option<String>,
    pub gtin: Option<String>,
    pub shipping_label: Option<String>,
    pub google_product_category: Option<String>,
    pub multipack: Option<String>,
    pub display_ads_id: Option<String>,
    pub display_ads_value: Option<f64>,
    pub id: Option<String>,
    pub shipping_width: Option<ProductShippingDimension>,
    pub pattern: Option<String>,
    pub unit_pricing_measure: Option<ProductUnitPricingMeasure>,
    pub custom_attributes: Option<Vec<ProductCustomAttribute>>,
    pub availability_date: Option<String>,
    pub is_bundle: Option<bool>,
    pub channel: Option<String>,
    pub description: Option<String>,
    pub warnings: Option<Vec<ErrorType>>,
    pub price: Option<Price>,
    pub shipping_length: Option<ProductShippingDimension>,
    pub display_ads_link: Option<String>,
    pub validated_destinations: Option<Vec<String>>,
    pub custom_groups: Option<Vec<ProductCustomGroup>>,
    pub loyalty_points: Option<LoyaltyPoints>,
    pub unit_pricing_base_measure: Option<ProductUnitPricingBaseMeasure>,
    pub shipping: Option<Vec<ProductShipping>>,
    pub shipping_weight: Option<ProductShippingWeight>,
    pub image_link: Option<String>,
    pub display_ads_title: Option<String>,
}

Product data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete products (none)

custombatch products (none)

get products (response)

list products (none)

insert products (request|response)

Fields

adult

Set to true if the item is targeted towards adults.

color

Color of the item.

adwords_labels

Similar to adwords_grouping, but only works on CPC.

item_group_id

Shared identifier for all variants of the same product.

expiration_date

Date on which the item should expire, as specified upon insertion, in ISO 8601 format. The actual expiration date in Google Shopping is exposed in productstatuses as googleExpirationDate and might be earlier if expirationDate is too far in the future.

availability

Availability status of the item.

adwords_grouping

Used to group items in an arbitrary way. Only for CPA%, discouraged otherwise.

destinations

Specifies the intended destinations for the product.

target_country

The two-letter ISO 3166 country code for the item.

custom_label4

Custom label 4 for custom grouping of items in a Shopping campaign.

title

Title of the item.

online_only

Whether an item is available for purchase only online.

mobile_link

Link to a mobile-optimized version of the landing page.

aspects

Specifies the intended aspects for the product.

shipping_height

Height of the item for shipping.

content_language

The two-letter ISO 639-1 language code for the item.

mpn

Manufacturer Part Number (MPN) of the item.

sale_price_effective_date

Date range during which the item is on sale (see product feed specifications).

brand

Brand of the item.

material

The material of which the item is made.

link

URL directly linking to your item's page on your website.

adwords_redirect

Allows advertisers to override the item URL when the product is shown within the context of Product Ads.

energy_efficiency_class

The energy efficiency class as defined in EU directive 2010/30/EU.

size_system

System in which the size is specified. Recommended for apparel items.

size_type

The cut of the item. Recommended for apparel items.

custom_label3

Custom label 3 for custom grouping of items in a Shopping campaign.

custom_label2

Custom label 2 for custom grouping of items in a Shopping campaign.

condition

Condition or state of the item.

custom_label0

Custom label 0 for custom grouping of items in a Shopping campaign.

kind

Identifies what kind of resource this is. Value: the fixed string "content#product".

installment

Number and amount of installments to pay for an item. Brazil only.

sizes

Size of the item.

gender

Target gender of the item.

taxes

Tax information.

identifier_exists

False when the item does not have unique product identifiers appropriate to its category, such as GTIN, MPN, and brand. Required according to the Unique Product Identifier Rules for all target countries except for Canada.

sale_price

Advertised sale price of the item.

product_type

Your category of the item (formatted as in product feeds specification).

display_ads_similar_ids

Advertiser-specified recommendations.

custom_label1

Custom label 1 for custom grouping of items in a Shopping campaign.

age_group

Target age group of the item.

additional_image_links

Additional URLs of images of the item.

offer_id

An identifier of the item.

gtin

Global Trade Item Number (GTIN) of the item.

shipping_label

The shipping label of the product, used to group product in account-level shipping rules.

google_product_category

Google's category of the item (see Google product taxonomy).

multipack

The number of identical products in a merchant-defined multipack.

display_ads_id

An identifier for an item for dynamic remarketing campaigns.

display_ads_value

Offer margin for dynamic remarketing campaigns.

id

The REST id of the product.

shipping_width

Width of the item for shipping.

pattern

The item's pattern (e.g. polka dots).

unit_pricing_measure

The measure and dimension of an item.

custom_attributes

A list of custom (merchant-provided) attributes. It can also be used for submitting any attribute of the feed specification in its generic form (e.g., { "name": "size type", "type": "text", "value": "regular" }). This is useful for submitting attributes not explicitly exposed by the API.

availability_date

The day a pre-ordered product becomes available for delivery, in ISO 8601 format.

is_bundle

Whether the item is a merchant-defined bundle. A bundle is a custom grouping of different products sold by a merchant for a single price.

channel

The item's channel (online or local).

description

Description of the item.

warnings

Read-only warnings.

price

Price of the item.

shipping_length

Length of the item for shipping.

display_ads_link

URL directly to your item's landing page for dynamic remarketing campaigns.

validated_destinations

The read-only list of intended destinations which passed validation.

custom_groups

A list of custom (merchant-provided) custom attribute groups.

loyalty_points

Loyalty points that users receive after purchasing the item. Japan only.

unit_pricing_base_measure

The preference of the denominator of the unit price.

shipping

Shipping rules.

shipping_weight

Weight of the item for shipping.

image_link

URL of an image of the item.

display_ads_title

Title of an item for dynamic remarketing campaigns.